% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute.R
\name{addClass}
\alias{addClass}
\title{Add a Class to an Object}
\usage{
addClass(x, className, first = TRUE, dbg = FALSE)
}
\arguments{
\item{x}{R object}

\item{className}{name of the class to be added to \code{x}}

\item{first}{if \code{TRUE} (default) the \code{className} is prepended to
the vector of existing class names, otherwise appended to the end of the
vector}

\item{dbg}{if \code{TRUE} (default), debug messages are shown.}
}
\value{
\code{x} with the \code{class} attributed extended by
  \code{className} (only if \code{x} did not yet inherit from
  \code{className})
}
\description{
Extend the class attribute by a given class name
}
