% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsLogical.R
\name{allAreIdentical}
\alias{allAreIdentical}
\title{Are all list elements identical to each other?}
\usage{
allAreIdentical(x)
}
\arguments{
\item{x}{a list}
}
\value{
\code{TRUE} if all elements in \code{x} are identical, otherwise 
  \code{FALSE}
}
\description{
Are all list elements identical to each other?
}
