% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsLogical.R
\name{almostEqual}
\alias{almostEqual}
\title{compare numerical vectors with tolerance}
\usage{
almostEqual(x, y, tolerance = 1e-12)
}
\arguments{
\item{x}{vector of numerical}

\item{y}{vector of numerical}

\item{tolerance}{tolerance, i.e. accepted difference between values in \code{x} and 
\code{y}. Default: 1e-12}
}
\value{
vector of logical
}
\description{
compare numerical vectors accepting a tolerance
}
