% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv.R
\name{argsCsv}
\alias{argsCsv}
\title{Language Abbreviation to "sep" and "dec" argument}
\usage{
argsCsv(lng = "de")
}
\arguments{
\item{lng}{"de" for German, "en" for Englisch}
}
\value{
list with elements \code{sep} (column separator), \code{dec} 
  (character used for decimal points)
}
\description{
Language Abbreviation to "sep" and "dec" argument
}
