% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrayToDataFrame.R
\name{arrayToDataFrame}
\alias{arrayToDataFrame}
\title{Convert Array to Data Frame}
\usage{
arrayToDataFrame(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{an array}

\item{name}{name to be given to the value column}
}
\value{
data frame with as many rows as there are elements in the input array
  \code{x}. The data frame has one column per dimension of \code{x}
  containing the dimension names and one additional column called according
  to \code{name} containing the array values.
}
\description{
Convert Array to Data Frame
}
\examples{
x <- array(1:24, dim = c(2, 3, 4), dimnames = list(
  c("a", "b"), 
  c("x", "y", "z"), 
  c("r", "s", "t", "u")
))

arrayToDataFrame(x)

}
