% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{asNoFactorDataFrame}
\alias{asNoFactorDataFrame}
\title{Convert to Data Frame without Factors}
\usage{
asNoFactorDataFrame(...)
}
\arguments{
\item{\dots}{passed to \code{\link{as.data.frame}}}
}
\description{
Use \code{\link{as.data.frame}} with \code{stringsAsFactors = FALSE}
}
\examples{
data_matrix <- matrix(LETTERS[1:6], nrow = 2)

# as.data.frame() by default converts character to factor
str(as.data.frame(data_matrix))

# asNoFactorDataFrame keeps character as character
str(asNoFactorDataFrame(data_matrix))

}
