% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{asRowList}
\alias{asRowList}
\title{Matrix to List of Matrix Rows}
\usage{
asRowList(x)
}
\arguments{
\item{x}{matrix}
}
\value{
list with as many elements as there are rows in \code{x} and each
  element representing one row
}
\description{
Matrix to List of Matrix Rows
}
\examples{
x <- matrix(1:12, nrow = 3)

row_list <- asRowList(x)

for (i in 1:nrow(x)) print(identical(row_list[[i]], x[i, ]))

}
