% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertFinalSlash.R
\name{assertFinalSlash}
\alias{assertFinalSlash}
\title{Make Sure that Strings End With Slash}
\usage{
assertFinalSlash(x, method = 1L)
}
\arguments{
\item{x}{vector of character}

\item{method}{integer value specifying the implementation method. \code{1}
(default): Find strings without ending slash and append slash to these
strings. \code{2}: Remove one or more slashes at the end and append slash
to all strings. \code{3}: Append slash to all strings and replace multiple
occurrences of slash at the end with one slash. Method 1 is the fastest but
does not replace multiple trailing slashes with only one trailing slash 
(see examples).}
}
\description{
Make Sure that Strings End With Slash
}
\examples{
assertFinalSlash(c("a", "b", "c"))
assertFinalSlash(c("a/", "b/", "c/"))
assertFinalSlash(c("a//", "b", "c/"))

# Use method 2 or 3 to replace multiple slashes with one slash
assertFinalSlash(c("a//", "b", "c/"), method = 2)
assertFinalSlash(c("a//", "b", "c/"), method = 3)
}
