% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{assignAll}
\alias{assignAll}
\title{Call assign for each List Element}
\usage{
assignAll(x, pos = 1, ...)
}
\arguments{
\item{x}{list of \code{key = value} assignments}

\item{pos}{passed to \code{assign},}

\item{\dots}{further arguments passed to \code{assign}}
}
\description{
Provide all assignments of a list as objects in the environment
}
