% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_argument_defaults.R
\name{assignArgumentDefaults}
\alias{assignArgumentDefaults}
\title{Assign Argument Default Values in the Global Environment}
\usage{
assignArgumentDefaults(FUN)
}
\arguments{
\item{FUN}{a function (a closure or a primitive). If \code{FUN} is a
character string then the function with that name is found and used.}
}
\description{
Assign Argument Default Values in the Global Environment
}
\examples{
# Assign the default values of breakInSequence() in the global environment
assignArgumentDefaults(kwb.utils::breakInSequence)

# The argument "expectedDiff" is now in the global environment
ls()

# Its value is 1 which is the default value defined in the function definition
expectedDiff

}
