% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{assignGlobally}
\alias{assignGlobally}
\title{assignGlobally}
\usage{
assignGlobally(x, value)
}
\arguments{
\item{x}{name of variable}

\item{value}{value of variable}
}
\description{
assign variable in .GlobalEnv
}
