% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFunctionAssignObjects.R
\name{assignObjects}
\alias{assignObjects}
\title{Assign Package Functions to the Global Environment}
\usage{
assignObjects()
}
\description{
This function provides all (also non-exported) function definitions of this 
package in the Global environment. This is useful for debugging the code
of a function that calls non-exported functions.
}
\examples{
\dontrun{
# Provide all functions of kwb.utils in the global environment
assignObjects()
}
}
