% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{assignPackageObjects}
\alias{assignPackageObjects}
\title{Assign all Package Objects to the Global Environment}
\usage{
assignPackageObjects(package)
}
\arguments{
\item{package}{package name}
}
\description{
Assign all Package Objects to the Global Environment
}
\examples{
assignPackageObjects("kwb.utils")

}
