\name{atLeastOneRowIn}
\alias{atLeastOneRowIn}
\title{at least one row in data frame}
\description{returns TRUE if data frame has at least one row, else FALSE}
\usage{atLeastOneRowIn(dframe)}
\arguments{
  \item{dframe}{data frame}
}



\author{Hauke Sonnenberg}





