% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{breakInSequence}
\alias{breakInSequence}
\title{Find "Breaks" in a Sequence of Numbers}
\usage{
breakInSequence(x, expectedDiff = 1)
}
\arguments{
\item{x}{vector of numeric}

\item{expectedDiff}{expected difference between elements in x. A bigger
difference is recognised as a break. Default: 1}
}
\value{
index of elements after which a break occurs or integer(0) if no
  break occurs at all
}
\description{
Find "Breaks" in a Sequence of Numbers
}
