% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{callWithStringsAsFactors}
\alias{callWithStringsAsFactors}
\title{Call a Function with Option "stringsAsFactors" set temporarily}
\usage{
callWithStringsAsFactors(stringsAsFactors, FUN, ...)
}
\arguments{
\item{stringsAsFactors}{TRUE or FALSE. Before calling \code{FUN} the option
"stringsAsFactors" is set to the value given here. After the function call
the option is reset to what it was before.}

\item{FUN}{function to be called}

\item{\dots}{arguments passed to \code{FUN}}
}
\value{
This function returns what \code{FUN} returns when called with the
  arguments given in \code{...}
}
\description{
Set the option "stringsAsFactors", run a function and reset the option.
}
\examples{
option.bak <- getOption("stringsAsFactors")
  
d1 <- callWithStringsAsFactors(
  TRUE,
  rbind,
  data.frame(id = 1, name = "Peter"),
  data.frame(id = 2, name = "Paul"),
  data.frame(id = 3, name = "Mary")
)
  
d2 <- callWithStringsAsFactors(
  FALSE,
  rbind,
  data.frame(id = 1, name = "Peter"),
  data.frame(id = 2, name = "Paul"),
  data.frame(id = 3, name = "Mary")
)
  
str(d1)
str(d2)
  
# The option "stringsAsFactors" has not changed!
stopifnot(option.bak == getOption("stringsAsFactors"))
  
}
