% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catChanges.R
\name{catChanges}
\alias{catChanges}
\title{Print Differences Between Two Vectors}
\usage{
catChanges(x, y)
}
\arguments{
\item{x}{vector}

\item{y}{vector}
}
\description{
Print Differences Between Two Vectors
}
\examples{
catChanges(c(1, 2, 3), c(1, 22, 3))
}
