% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catChangesIf.R
\name{catChangesIf}
\alias{catChangesIf}
\title{Print Differences Between Two Vectors if Condition is Met}
\usage{
catChangesIf(dbg, x, y)
}
\arguments{
\item{dbg}{if \code{TRUE} \code{\link{catChanges}} is called on \code{x} and
\code{y}}

\item{x}{vector}

\item{y}{vector}
}
\description{
Print Differences Between Two Vectors if Condition is Met
}
