% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catchWarning.R
\name{catchWarning}
\alias{catchWarning}
\title{Return Warning Message in Attribute}
\usage{
catchWarning(expr, dbg = TRUE)
}
\arguments{
\item{expr}{expression to be evaluated}

\item{dbg}{if \code{TRUE} (the default) the warning text is also printed on
the console, otherwise the warning message is suppressed.}
}
\description{
This function evaluates an expression and returns, if warnings occurred, the
warning message in the attribute "warningMessage" of the returned object.
}
\examples{
catchWarning(as.numeric("1.23"))
result <- catchWarning(as.numeric("x"))
result <- catchWarning(as.numeric("x"), dbg = FALSE)
str(result)
}
