\name{catIf}
\alias{catIf}
\title{call cat if condition is met}
\description{call cat if condition is met}
\usage{catIf(condition, ...)}
\arguments{
  \item{condition}{if TRUE, cat is called, else not}
  \item{\dots}{arguments passed to cat}
}



\author{Hauke Sonnenberg}





