% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{catLines}
\alias{catLines}
\title{Print Character Vector to the Console}
\usage{
catLines(x)
}
\arguments{
\item{x}{vector of character representing text lines to be printed}
}
\description{
Call cat on character vector, pasted with collapse = <new line>
}
