% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{catNewLineIf}
\alias{catNewLineIf}
\title{Print New Line Character to the Console if Condition is Met}
\usage{
catNewLineIf(condition)
}
\arguments{
\item{condition}{if \code{TRUE} the new line is printed else not}
}
\value{
Returns the condition, invisibly so that it can be reused
}
\description{
Print New Line Character to the Console if Condition is Met
}
