% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipMatrix.R
\name{clipMatrix}
\alias{clipMatrix}
\title{Remove Leading/Trailing Empty Rows/Columns from Matrix}
\usage{
clipMatrix(m)
}
\arguments{
\item{m}{a matrix}
}
\value{
\code{m} with leading trailing empty rows (full of \code{NA}) removed
}
\description{
Remove Leading/Trailing Empty Rows/Columns from Matrix
}
