% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{cmdLinePath}
\alias{cmdLinePath}
\title{Path in Quotes for Usage in Command Line}
\usage{
cmdLinePath(x)
}
\arguments{
\item{x}{path to be quoted}
}
\description{
Set the given path in quotes so that in can be used in a command line
#'
}
