% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{collapsed}
\alias{collapsed}
\title{shortcut to paste(x, collapse = collapse)}
\usage{
collapsed(x, collapse = " ")
}
\arguments{
\item{x}{vector of character}

\item{collapse}{character string to separate the elements in x (passed to 
\code{\link[base]{paste}})}
}
\description{
This is just a shortcut to paste(x, collapse = collapse)
}
