% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsStat.R
\name{colMaxima}
\alias{colMaxima}
\title{Columnwise Maxima}
\usage{
colMaxima(dataFrame, na.rm = FALSE)
}
\arguments{
\item{dataFrame}{data frame of which to calculate columnwise maxima}

\item{na.rm}{passed to the \code{max} function}
}
\description{
Calculate the maxima within each column
}
