% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsStat.R
\name{colMinima}
\alias{colMinima}
\title{Columnwise Minima}
\usage{
colMinima(dataFrame, na.rm = FALSE)
}
\arguments{
\item{dataFrame}{data frame of which to calculate columnwise minima}

\item{na.rm}{passed to the \code{min} function}
}
\description{
Calculate the minima within each column
}
