% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{colNaNumbers}
\alias{colNaNumbers}
\title{Columnwise Number of NA}
\usage{
colNaNumbers(dataFrame)
}
\arguments{
\item{dataFrame}{data frame of which to calculate columnwise NA values}
}
\description{
Calculate the number of NA values within each column
}
