% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{colStatisticOneFunction}
\alias{colStatisticOneFunction}
\title{Apply Function to All Columns}
\usage{
colStatisticOneFunction(dataFrame, FUN, na.rm = FALSE)
}
\arguments{
\item{dataFrame}{a data frame of which statistics are to be calculated}

\item{FUN}{statistical function to be applied on each column of dataFrame
possible values: "sum", "mean", "min", "max", "number.na" (number of NA 
values), "length" (number of values)}

\item{na.rm}{if TRUE, NA values are removed before applying the statistical function}
}
\description{
applies a statistical function to all columns of a data frame
}
