% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{combineAlternatingly}
\alias{combineAlternatingly}
\title{Combine Two Vectors Alternatingly}
\usage{
combineAlternatingly(a, b)
}
\arguments{
\item{a}{first vector}

\item{b}{second vector}
}
\value{
vector \code{x} with all \code{x[c(1, 3, 5, ...)] == a[c(1, 2, 3,
  ...)]} and all \code{x[c(2, 4, 6, ...)] == b[c(1, 2, 3, ...)]}
}
\description{
Combine Two Vectors Alternatingly
}
\examples{
a <- paste0("a", 1:5)
b <- paste0("b", 1:5)

combineAlternatingly(a, b)
combineAlternatingly(b, a)

combineAlternatingly(list(a = 1, b = 2), list(c = 3, d = 4))

}
