% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonNames.R
\name{commonNames}
\alias{commonNames}
\title{Common Names in two Objects with Names Attribute}
\usage{
commonNames(x, y)
}
\arguments{
\item{x}{object with names attribute, e.g. data.frame, named vector}

\item{y}{object with names attribute, e.g. data.frame, named vector}
}
\value{
vector of names occurring in both \code{x} and \code{y}
}
\description{
Common Names in two Objects with Names Attribute
}
\examples{
x <- data.frame(a = 1:2, b = 2:3)
y <- c(a = 1, c = 2)
}
