% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{compareSets}
\alias{compareSets}
\title{Compare the values in two vectors}
\usage{
compareSets(
  x,
  y,
  subject = "Values",
  xname = deparse(substitute(x)),
  yname = deparse(substitute(y))
)
}
\arguments{
\item{x}{first vector}

\item{y}{second vector}

\item{subject}{name of objects to be compared that will appear in the 
message. Default: \code{"Values"}.}

\item{xname}{optional name of the first vector that will be used in the
message}

\item{yname}{optional name of the second vector that will be used in the
message}
}
\description{
Prints the result of comparing the values of two vectors with each other 
(which values are in the first vector but not in the second one and which 
values are in the second vector but not in the first one) on the screen.
}
\examples{
compareSets(1:10, 3:13)
compareSets(1:10, 3:13, "numbers", "set 1", "set 2")

}
