\name{containsNulString}
\alias{containsNulString}
\title{containsNulString}
\description{check for nul string in file}
\usage{containsNulString(filepath)}
\arguments{
  \item{filepath}{
}
}

\value{TRUE if first two bytes of file are FF FE, else FALSE}

\author{Hauke Sonnenberg}





