% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyAttributes.R
\name{copyAttributes}
\alias{copyAttributes}
\title{Copy Attributes Between Two Objects}
\usage{
copyAttributes(x, y, attrNames)
}
\arguments{
\item{x}{object to which to copy attributes of y}

\item{y}{object from which to copy attributes to x}

\item{attrNames}{vector of character containing the names of attributes in 
\code{y} to be copied to \code{x}}
}
\description{
Copy Attributes Between Two Objects
}
\examples{
x <- structure(1, a = 2, b = 3)
y <- structure(2, c = 4)
copyAttributes(x, y, "c")
}
