% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{copyDirectoryStructure}
\alias{copyDirectoryStructure}
\title{Copy Directory Structure}
\usage{
copyDirectoryStructure(
  sourcedir,
  targetdir,
  excludePattern = "^$",
  recursive = TRUE,
  dbg = TRUE
)
}
\arguments{
\item{sourcedir}{path to the source directory}

\item{targetdir}{path to the target directory}

\item{excludePattern}{pattern matching directory names to be excluded from
the copying process}

\item{recursive}{if TRUE (default) the full tree of directories and
subdirectories is copied, otherwise only the top-level directories}

\item{dbg}{if TRUE (default) debug messages are shown}
}
\value{
This function invisibly returns a vector of character containing the
  full paths of the directories that were created.
}
\description{
Copy the full directory structure from a source directory to a target 
directory
}
