% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{copyListElements}
\alias{copyListElements}
\title{Copy list elements into a list of lists}
\usage{
copyListElements(x, y, name)
}
\arguments{
\item{x}{list of lists}

\item{y}{list of elements}

\item{name}{name of target list element}
}
\value{
\code{x} with each sublist being extended by list element \code{name}
having been taken from \code{y}
}
\description{
Copy list elements into a list of lists
}
\examples{
x <- list(list(a = 1), list(b = 2), list(c = 3))
y <- list("b1", "b2", "b3")
str(copyListElements(x, y, "b"))
}
