% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{countNaInColumn}
\alias{countNaInColumn}
\title{Count NA in one Column of a Data Frame}
\usage{
countNaInColumn(data, column)
}
\arguments{
\item{data}{data frame}

\item{column}{column name}
}
\value{
number of \code{NA} in \code{column} of \code{data}
}
\description{
Count NA in one Column of a Data Frame
}
