% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{countOrSum}
\alias{countOrSum}
\title{Count or Sum Up Values Within Groups of rows}
\usage{
countOrSum(x, by = NULL, sum.up = NULL)
}
\arguments{
\item{x}{data frame}

\item{by}{vector of names of columns in \code{x} to be grouped by}

\item{sum.up}{name of column in \code{x} containing numeric values to be
summed up. If \code{NULL} (default) rows within groups are counted instead
of summing up values within each group}
}
\value{
object of class \code{xtabs} with as many dimensions as there are
  values in \code{by}
}
\description{
Count or Sum Up Values Within Groups of rows
}
\examples{
# Create a data frame with example data
x <- data.frame(
  Group = rep(c("A", "B", "C"), 4),
  Even = rep(c(FALSE, TRUE), 6),
  Value = seq_len(12)
)

# Count the number of rows for each group
countOrSum(x, "Group")
countOrSum(x, c("Group", "Even"))

# Sum up the values in column "Value" for each group
countOrSum(x, "Group", sum.up = "Value")
countOrSum(x, c("Group", "Even"), sum.up = "Value")
}
