% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createAccessor.R
\name{createAccessor}
\alias{createAccessor}
\title{Create Function to Safely Access Data Frame Columns or List Elements}
\usage{
createAccessor(x)
}
\arguments{
\item{x}{a data frame or a list}
}
\description{
Create Function to Safely Access Data Frame Columns or List Elements
}
\examples{
getcol <- createAccessor(data.frame(a = 1:2, b = 2:3))
getcol("a")
getcol("b")
getcol(c("b", "a"))
# getcol("c") # error with info about existing columns

}
