\name{createDirAndReturnPath}
\alias{createDirAndReturnPath}
\title{create directory if it does not exist}
\description{create directory if it does not exist}
\usage{createDirAndReturnPath(dir.to.create, dbg = TRUE, confirm = FALSE)}
\arguments{
  \item{dir.to.create}{
}
  \item{dbg}{
}
  \item{confirm}{
}
}



\author{Hauke Sonnenberg}





