% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDirectories.R
\name{createDirectories}
\alias{createDirectories}
\title{Create Directories}
\usage{
createDirectories(paths, ...)
}
\arguments{
\item{paths}{paths to directories to be created}

\item{\dots}{further arguments passed to \code{\link{createDirectory}}}
}
\value{
paths to created directories
}
\description{
Create Directories
}
