% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFunctionAssignObjects.R
\name{createFunctionAssignObjects}
\alias{createFunctionAssignObjects}
\title{Create Function to Provide Package Objects in Global Environment}
\usage{
createFunctionAssignObjects(package)
}
\arguments{
\item{package}{name of package for which to create a shortcut to
\code{assignPackageObjects(package = package)}}
}
\value{
This function returns a function that internally calls
  \code{assignPackageObjects(package = package)}
}
\description{
Create Function to Provide Package Objects in Global Environment
}
\examples{
provideFunctions <- createFunctionAssignObjects("kwb.utils")
\dontrun{
provideFunctions()
}
}
