% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFunctionExtdataFile.R
\name{createFunctionExtdataFile}
\alias{createFunctionExtdataFile}
\title{"Create Function to Compose Path in Package}
\usage{
createFunctionExtdataFile(package)
}
\arguments{
\item{package}{name of package for which to create a shortcut to
\code{system.file("extdata", ..., package = package)}}
}
\value{
This function returns a function that internally calls
  \code{system.file("extdata", ..., package = package")}
}
\description{
"Create Function to Compose Path in Package
}
\examples{
extdataFile <- createFunctionExtdataFile("kwb.utils")
dir(extdataFile())
}
