% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{createIdAlong}
\alias{createIdAlong}
\title{Create Unique Identifiers along Input Vector}
\usage{
createIdAlong(x, base_name = NULL)
}
\arguments{
\item{x}{vector of objects for which to create identifiers}

\item{base_name}{base string to which the hexadecimal suffixes are to be
appended. By default the name of the object \code{x}, with "s" at the
end removed, is used. For example, passing a vector called "files" to this
function results in ids "file_01", "file_02", ...}
}
\value{
vector of character as long as \code{x}
}
\description{
This function helps you create strings that can be used as identifiers.
}
\examples{
# Create ids for 32 numbers
createIdAlong(1:32, "number")
# Take base name from x, with the plural's "s" removed
numbers <- 1:32
createIdAlong(numbers)
}
