% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{createPasswordFile}
\alias{createPasswordFile}
\title{Create Encrypted Password File for Account}
\usage{
createPasswordFile(account, keyFile, passwordFile, password = NULL)
}
\arguments{
\item{account}{name of account the user is asked to enter the password for}

\item{keyFile}{path to the file containing the encryption/decryption key}

\item{passwordFile}{path to the password file}

\item{password}{password for account. If \code{NULL} (default) the user will
be asked to enter the password on the console}
}
\description{
Create Encrypted Password File for Account
}
