% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStorage.R
\name{createStorage}
\alias{createStorage}
\title{Create Storage Object to Save/Load .rds or .RData files}
\usage{
createStorage(path, type = "rds")
}
\arguments{
\item{path}{path to the folder in which to store the RDS files. If the folder 
does not exist it is attempted to be created.}

\item{type}{one of \code{"rds"}, \code{"RData"}, specifying the format used
to store R objects.}
}
\value{
list with "member functions" \code{list()}, \code{save()}, 
  \code{load()}, \code{remove()}, and variables \code{path}, \code{type}
}
\description{
This function returns an object that implements the methods \code{list()}, 
\code{save()}, \code{load()}, \code{remove()} that list, store, load or 
remove, respectively .rds or .RData files. The object is associated to a 
folder that is given to this function.
}
\examples{
# Create an empty test folder
path <- file.path(tempdir(), "test")
dir.create(path)

# Create a storage object pointing to a temporary test folder
storage <- kwb.utils::createStorage(path)

# At the beginning, the storage is empty
storage$list()

# Store objects
apple <- 5
storage$save(apple, banana = list("Hello", "World"))

\dontrun{
# List the objects
storage$list()

# Retrieve an object
storage$load("banana")

# Remove objects
storage$remove("apple")
storage$remove("banana")

# The storage is empty again
storage$list()
}
}
