\name{csvTextToDataFrame}
\alias{csvTextToDataFrame}
\title{csvTextToDataFrame}
\description{csvTextToDataFrame}
\usage{csvTextToDataFrame(text, ...)}
\arguments{
  \item{text}{character vector representing lines of comma separated values}
  \item{\dots}{arguments passed to read.table}
}



\author{Hauke Sonnenberg}





