% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_decode.R
\name{decode}
\alias{decode}
\title{Decode a Vector of Character}
\usage{
decode(x)
}
\arguments{
\item{x}{vector of encoded character strings as returned by 
\code{\link{encode}}}
}
\value{
vector of decoded character strings
}
\description{
For an example see \code{\link{encode}}.
}
