% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsMain.R
\name{defaultIfNA}
\alias{defaultIfNA}
\title{default value if object is NA}
\usage{
defaultIfNA(x, default, count = FALSE)
}
\arguments{
\item{x}{vector possibly containing NA values}

\item{default}{default value that is returned if \code{x} is NA}

\item{count}{if \code{TRUE} (the default is \code{FALSE}) the number of replaced
values is returned in the attributes \code{count}}
}
\value{
\code{x} if \code{x} is not NA and \code{default} otherwise
}
\description{
return the given object or a default value if the object is NA
}
\examples{
  defaultIfNA(NA, "default") # returns the default value
  defaultIfNA("actual", "default") # returns the "actual" value
  

}
\seealso{
\code{\link{defaultIfNULL}, \link{defaultIfZero}}
}
