% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsMain.R
\name{defaultIfNULL}
\alias{defaultIfNULL}
\title{default value if object is NULL}
\usage{
defaultIfNULL(x, default, count = FALSE)
}
\arguments{
\item{x}{R object to be checked for NULL}

\item{default}{default value that is returned if \code{x} is NULL}

\item{count}{if \code{TRUE} (the default is \code{FALSE}) the number of replaced
values is returned in the attributes \code{count}}
}
\value{
\code{x} if \code{x} is not NULL and \code{default} otherwise
}
\description{
return the given object or a default value if the object is NULL
}
\examples{
  defaultIfNULL(NULL, "default") # returns the default value
  defaultIfNULL("actual", "default") # returns the "actual" value
  

}
\seealso{
\code{\link{defaultIfNA}, \link{defaultIfZero}}
}
