% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{defaultIfZero}
\alias{defaultIfZero}
\title{Default Value if Object is 0 (zero)}
\usage{
defaultIfZero(x, default, count = FALSE)
}
\arguments{
\item{x}{vector possibly containing zeroes}

\item{default}{value to be used instead of zero}

\item{count}{if \code{TRUE} (the default is \code{FALSE}) the number of
replaced values is returned in the attributes \code{count}}
}
\value{
\code{x} if \code{x} is not 0 and \code{default} otherwise. The
  returned object has an attribute \code{count} containing the number of
  values that have been set to the default value.
}
\description{
Return the given object or a default value if the object is 0 (zero)
}
\examples{
defaultIfZero(c(1, 2, 0, 4, 5, 0, 6), NA) # returns the default value (NA) 
(out <- defaultIfZero(1:6, NA, count = TRUE)) # returns the "actual" values

# The number of values that were zero is returned in the attribute "count" 
attr(out, "count")
  
}
\seealso{
\code{\link{defaultIfNA}, \link{defaultIfNULL}}
}
