% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsMain.R
\name{defaultLevels}
\alias{defaultLevels}
\title{Default factor levels}
\usage{
defaultLevels(x, step = 1)
}
\arguments{
\item{x}{numeric or character vector}

\item{step}{step to be used in sequence generation}
}
\value{
if \code{x} is numeric the sequence between the lowest and highest value
  in \code{x} with the given \code{step} is returned. If \code{x} is a
  vector of character strings, the sorted unique values are returned.
}
\description{
If x is numeric and all values are whole numbers use the sequence between
  the smallest and the highest value
}
\examples{
  defaultLevels(c(1, 3, 4, 5, 4))
  defaultLevels(c(1920, 1950, 1970), step = 10)
  

}
