% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{desktop}
\alias{desktop}
\title{Path to Your Desktop}
\usage{
desktop(osType = .OStype())
}
\arguments{
\item{osType}{Optional. Type of operating system, one of \code{"unix"},
\code{"windows"}}
}
\value{
character string representing the path to your desktop
}
\description{
Get the path to your desktop
}
