% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{.defaultIf}
\alias{.defaultIf}
\title{Return a Default Value for Values Meeting a Condition}
\usage{
.defaultIf(FUN.test, x, default, count = FALSE)
}
\arguments{
\item{count}{if \code{TRUE} (the default is \code{FALSE}) the number of
replaced values is returned in the attributes \code{count}}
}
\description{
Return a Default Value for Values Meeting a Condition
}
