% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.guessSeparator.1}
\alias{.guessSeparator.1}
\title{Guess Column Separator (Version 1)}
\usage{
.guessSeparator.1(textlines, separators, comment.char = "#")
}
\description{
Guess Column Separator (Version 1)
}
