% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.guessSeparator.2}
\alias{.guessSeparator.2}
\title{Guess Column Separator (Version 2)}
\usage{
.guessSeparator.2(textlines, separators)
}
\arguments{
\item{textlines}{vector of character representing the first lines of a file}

\item{separators}{vector of possible column separator characters the file is
to be checked for}
}
\description{
Guess Column Separator (Version 2)
}
