% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{.log}
\alias{.log}
\title{Write Log Message to Console}
\usage{
.log(...)
}
\arguments{
\item{\dots}{parts of the message that will be passed to \code{\link{cat}}}
}
\description{
Write Log Message to Console
}
