% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{.logline}
\alias{.logline}
\title{Write Log Message and New Line Character to Console}
\usage{
.logline(...)
}
\arguments{
\item{\dots}{parts of the message that will be passed to \code{\link{cat}}}
}
\description{
Write Log Message and New Line Character to Console
}
