% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{.logok}
\alias{.logok}
\title{Write "ok" and New Line Character to Console}
\usage{
.logok(dbg = TRUE)
}
\arguments{
\item{dbg}{logical to switch logging on (\code{dbg = TRUE}) or off (\code{dbg
= FALSE})}
}
\description{
Write "ok" and New Line Character to Console
}
