% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{.logstart}
\alias{.logstart}
\title{Write Log Message to Console if in Debugging Mode}
\usage{
.logstart(dbg = TRUE, ...)
}
\arguments{
\item{dbg}{logical to switch logging on (\code{dbg = TRUE}) or off (\code{dbg
= FALSE})}

\item{\dots}{parts of the message that will be passed to \code{\link{catIf}}}
}
\description{
Write Log Message to Console if in Debugging Mode
}
