% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{.OStype}
\alias{.OStype}
\title{see tools:::.OStype}
\usage{
.OStype()
}
\description{
This is a copy of tools:::.OStype. It is redefined here so that it can be
used within this package. R CMD build would complain if I used
tools:::.OStype!
}
