% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{dropUnusedFactorLevels}
\alias{dropUnusedFactorLevels}
\title{Drop Unused Factor Levels in all Factor Columns}
\usage{
dropUnusedFactorLevels(data, dbg = TRUE)
}
\arguments{
\item{data}{data frame in which to remove unused levels in all columns that 
are factors}

\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\value{
\code{data} with unused factors removed in all columns being factors
}
\description{
Drop Unused Factor Levels in all Factor Columns
}
\examples{
# Create an example data frame with two factor columns
data <- data.frame(
  id = 1:3, 
  factor_1 = factor(c("a", "b", "a"), levels = c("a", "b", "c")),
  factor_2 = factor(c("x", "x", "y"), levels = c("x", "y", "z")),
  no_factor = c("A", "B", "C"),
  stringsAsFactors = FALSE
)

# Review the structure of the data frame
str(data)

# Review the structure of the data frame with unused factors removed.
str(dropUnusedFactorLevels(data))

}
