% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_decode.R
\name{encode}
\alias{encode}
\title{Encode a Vector of Character}
\usage{
encode(x, level = 1, chars = printable_chars(level))
}
\arguments{
\item{x}{vector of character}

\item{level}{one of 1, 2, or 3 defining which character set to use for 
encoding, see \code{\link{printable_chars}}.}

\item{chars}{vector of characters to be used for encoding}
}
\value{
vector of character as long as \code{x} with each element containing
  the encoded version of the corresponding element in \code{x}. The returned
  vector has an attribute \code{codes} being a named vector. This vector 
  contains the unique values in \code{x} as elements. Each element is named
  by the code that was used to encode the corresponding element.
}
\description{
Encode a Vector of Character
}
\examples{
x <- c("very-long-word", "very-long-word", "very-very-long-word")
encoded <- encode(x)
encoded
decode(encoded)

}
