% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{excludeNULL}
\alias{excludeNULL}
\title{Exclude all NULL Entries from a List}
\usage{
excludeNULL(x, dbg = TRUE)
}
\arguments{
\item{x}{a list}

\item{dbg}{if \code{TRUE} (default) a message is shown if elements are 
removed}
}
\value{
list \code{x} with all \code{NULL} entries excluded
}
\description{
Exclude all NULL Entries from a List
}
\examples{
L <- list(a = 1, b = NULL, c = "three")
L

excludeNULL(L)

}
