% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{expandGrid}
\alias{expandGrid}
\title{Wrapper around expand.grid}
\usage{
expandGrid(..., stringsAsFactors = FALSE)
}
\arguments{
\item{...}{arguments passed to \code{\link[base]{expand.grid}}, but in 
reversed order}

\item{stringsAsFactors}{passed to \code{\link[base]{expand.grid}}}
}
\description{
Same as \code{\link[base]{expand.grid}} but with \code{stringsAsFactors =
FALSE} by default and with the values of the first argument being changed
last, not first.
}
\examples{
persons <- c("Peter", "Paul", "Mary")
fruits <- c("apple", "pear")

# With expand.grid() the values of the first argument change first...
(grid_1 <- expand.grid(person = persons, fruit = fruits))

#... with expandGrid() they change last.
grid_2 <- expandGrid(person = persons, fruit = fruits)

# With expand.grid() character strings are converted to factors by default...
str(grid_1)

# ... with expandGrid() character strings are not converted by default.
str(grid_2)
}
