\name{extendLimits}
\alias{extendLimits}
\title{extendLimits}
\description{extendLimits}
\usage{extendLimits(limits, left = 0.05, right = left, absolute = FALSE)}
\arguments{
  \item{limits}{vector of two elements as e.g. used for xlim or ylim}
  \item{left}{percentage of limit range (\emph{absolute == FALSE}) or absolute value
(\emph{absolute == TRUE}) by which the left limit is extended to the left.}
  \item{right}{percentage of limit range (\emph{absolute == FALSE}) or absolute value 
(\emph{absolute == TRUE}) by which the right limit is extended to the
right.}
  \item{absolute}{Default: FALSE}
}



\author{Hauke Sonnenberg}





