% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{fileExtension}
\alias{fileExtension}
\title{Get File Extension of Full Paths(s)}
\usage{
fileExtension(x)
}
\arguments{
\item{x}{vector of file paths}
}
\description{
Get File Extension of Full Paths(s)
}
\examples{
# Define example paths
paths <- c("C:/example/file.csv", "file2.txt", "D:/e/f/ghi.jkl.zip")

# Get the file name extensions
fileExtension(paths)

# Empty string is returned for paths without file name extensions
fileExtension("C:/NEWS")
}
