% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findChanges.R
\name{findChanges}
\alias{findChanges}
\title{Find Value Changes in a Vector}
\usage{
findChanges(x)
}
\arguments{
\item{x}{vector of atomic mode (e.g. logical, numeric, character)}
}
\value{
data frame with one row more than there are value changes in
  \code{x}. If the value in \code{x} does not change from one index to the
  next, it is assumed to belong to the same index region. If the value
  changes, a new index region begins. In the result data frame each index
  region is given by \code{starts_at} and \code{ends_at} which are the
  indices of the first and last element, respectively, of each index region.
  The function returns \code{NULL} for input vectors \code{x} of length 0.
}
\description{
Finds the positions in a vector where the value changes an returns the "index
regions", i.e. the regions between indices \code{starts_at} and
\code{ends_at} whithin each of which the value does not change, i.e. 
\code{length(unique(x[starts_at:ends_at])) == 1L} is \code{TRUE}.
}
\examples{
findChanges(c(1,2,2,3,3,3))
findChanges(c("a", "a", "b", "c", "c", "d"))
findChanges(c(TRUE, TRUE, TRUE, FALSE, FALSE, TRUE, FALSE, FALSE))

}
