% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{finishAndShowPdf}
\alias{finishAndShowPdf}
\title{Close Device and Open PDF File in Viewer}
\usage{
finishAndShowPdf(PDF, which = grDevices::dev.cur(), ...)
}
\arguments{
\item{PDF}{full path to the PDF file to be opened in the PDF Viewer}

\item{which}{passed to \code{grDevices::dev.off}}

\item{\dots}{further arguments passed to \code{hsShowPdf}}
}
\description{
Close the PDF device (default: \code{dev.cur()}) and open the PDF file in a
PDF viewer
}
