\name{finishAndShowPdfIf}
\alias{finishAndShowPdfIf}
\title{finishAndShowPdfIf}
\description{finish and display pdf file if condition is met}
\usage{finishAndShowPdfIf(to.pdf, PDF, ...)}
\arguments{
  \item{to.pdf}{
}
  \item{PDF}{
}
  \item{\dots}{
}
}



\author{Hauke Sonnenberg}





