% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{firstPosixColumn}
\alias{firstPosixColumn}
\title{data/time column of data frame}
\usage{
firstPosixColumn(x)
}
\arguments{
\item{x}{data frame in which to find a column of class "POSIXt"}
}
\description{
data/time column of data frame
}
