\name{frenchToAscii}
\alias{frenchToAscii}
\title{French unicode letter to ASCII letter(s)}
\description{French unicode letter to ASCII letter(s)}
\usage{frenchToAscii()}

\value{list of ASCII characters (list elements) replacing unicode characters
(element names)}

\author{Hauke Sonnenberg}





