% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullWinPath.R
\name{fullWinPath}
\alias{fullWinPath}
\title{Windows-Compatible Expanded File Path}
\usage{
fullWinPath(x, doubleBackslashes = FALSE)
}
\arguments{
\item{x}{file path (with slashes or backslashes as separators)}

\item{doubleBackslashes}{if \code{TRUE} the backslashes in the Windows 
compatible path are duplicated each}
}
\value{
Windows compatible file path with backslashes (duplicated if
  \code{doubeBackslashes} is \code{TRUE})
}
\description{
Expanded file path with backslashes (each duplicated if 
\code{doubleBackslashes} is \code{TRUE}). Duplicated backslashes can be 
useful when being used in RMarkdown documents
}
\examples{
cat(fullWinPath("~/a/b"))
cat(fullWinPath("~/a/b", doubleBackslashes = TRUE))
}
