\name{generateKeyFile}
\alias{generateKeyFile}
\title{generate a decryption key file}
\description{generate a decryption key file}
\usage{generateKeyFile(target)}
\arguments{
  \item{target}{full path to the file to which the key shall be written}
}



\author{Hauke Sonnenberg}





