% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getAttribute}
\alias{getAttribute}
\title{safely get the attribute of an object}
\usage{
getAttribute(x, attributeName, do.stop = TRUE)
}
\arguments{
\item{x}{object from which to take the attribute}

\item{attributeName}{name of the attribute to be returned}

\item{do.stop}{if \code{TRUE} (default) an error is raised if the attribute does not
exist.}
}
\description{
safely get the attribute of an object. An error is given if the attribute
  does not exist (and do.stop = TRUE)
}
