\name{getByPositiveOrNegativeIndex}
\alias{getByPositiveOrNegativeIndex}
\title{getByPositiveOrNegativeIndex}
\description{get element from vector, counting from head or tail}
\usage{getByPositiveOrNegativeIndex(elements, index)}
\arguments{
  \item{elements}{vector of elements}
  \item{index}{positive or negative index(es) with absolute value between 1 and
length(\emph{elements})}
}

\value{element(s) out of \emph{elements} corresponding to the index(es) given in 
\emph{index}}

\author{Hauke Sonnenberg}





