% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{getElementLengths}
\alias{getElementLengths}
\title{Get the Lenghts of List Elements}
\usage{
getElementLengths(x)
}
\arguments{
\item{x}{a list}
}
\value{
vector of integer
}
\description{
Get the Lenghts of List Elements
}
\examples{
x <- list(a = 1:3, b = list(x = 1, y = 2), c = 1:100)
getElementLengths(x)

}
