% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getEvenNumbers}
\alias{getEvenNumbers}
\title{Get even Numbers out of a Vector of Integers}
\usage{
getEvenNumbers(x)
}
\arguments{
\item{x}{vector of integer}
}
\description{
Get even Numbers out of a Vector of Integers
}
