% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getFunctionName}
\alias{getFunctionName}
\title{get the name of a function}
\usage{
getFunctionName(FUN)
}
\arguments{
\item{FUN}{R object representing a function}
}
\description{
get the name of a function
}
