% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default.R
\name{getFunctionValueOrDefault}
\alias{getFunctionValueOrDefault}
\title{Function Value or Default if NA}
\usage{
getFunctionValueOrDefault(values, FUN, default, warningMessage = NA)
}
\arguments{
\item{values}{vector of values given to \code{FUN}}

\item{FUN}{function to which values are passed and which offers the argument
"na.rm"}

\item{default}{default value to be returned if all values are NA}

\item{warningMessage}{Warning message given if the default was taken}
}
\description{
Take the function value or a default value if the function value is NA
}
