% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getGlobally}
\alias{getGlobally}
\title{getGlobally}
\usage{
getGlobally(x, default = NULL, create.if.not.existing = TRUE)
}
\arguments{
\item{x}{name of variable}

\item{default}{default value to which the variable is assigned (if
create.if.not.existing = TRUE) in case that it does not yet exist. Default:
NULL}

\item{create.if.not.existing}{if TRUE and if the variable does not yet exist,
it is created and initialised with the value given in \code{default}.
Default: TRUE}
}
\description{
gat variable from .GlobalEnv
}
