% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsParse.R
\name{getKeywordPositions}
\alias{getKeywordPositions}
\title{localise keywords in data frame}
\usage{
getKeywordPositions(dataFrame, keywords, asDataFrame = TRUE)
}
\arguments{
\item{dataFrame}{data frame or matrix in which to search for given keywords}

\item{keywords}{(list of) keywords to be looked for in \emph{data frame}}

\item{asDataFrame}{if TRUE (default), a data frame is returned, otherwise a matrix}
}
\value{
data frame (if \emph{asDataFrame} = TRUE) or matrix with one column per 
  keyword that was given in \emph{keywords}. The first row contains the row
  numbers and the second row contains the column numbers, respectively, of
  the fields in \emph{dataFrame} in which the corresponding keywords were
  found.
}
\description{
localise keywords in data frame
}
