% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getOddNumbers}
\alias{getOddNumbers}
\title{Get odd Numbers out of a Vector of Integers}
\usage{
getOddNumbers(x)
}
\arguments{
\item{x}{vector of integer}
}
\description{
Get odd Numbers out of a Vector of Integers
}
