\name{getPassword}
\alias{getPassword}
\title{get encrypted password from file using key}
\description{get encrypted password from file using key}
\usage{getPassword(passwordFile, keyFile)}
\arguments{
  \item{passwordFile}{
}
  \item{keyFile}{
}
}

\value{NA if no password is stored  }

\author{Hauke Sonnenberg}





