% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{guessSeparator}
\alias{guessSeparator}
\title{Guess Column Separator Used in File}
\usage{
guessSeparator(csvFile, n = 10, separators = c(";", ",", "\\t"))
}
\arguments{
\item{csvFile}{full path to text file containing 'comma separated values'}

\item{n}{number of first lines in the file to be looked at}

\item{separators}{vector of possible column separator characters the file is
to be checked for}
}
\description{
Guess Column Separator Used in File
}
