% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasFinalSlash.R
\name{hasFinalSlash}
\alias{hasFinalSlash}
\title{Does a String End with Slash?}
\usage{
hasFinalSlash(x)
}
\arguments{
\item{x}{vector of character}
}
\value{
vector of boolean being \code{TRUE} at indices where the elements in
  \code{x} end with a slash ("/")
}
\description{
Does a String End with Slash?
}
\examples{
(is_directory <- hasFinalSlash(c("a", "b/", "c", "d/")))
}
