% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hasZeroLength.R
\name{hasZeroLength}
\alias{hasZeroLength}
\title{Does an Object Have a Length of Zero?}
\usage{
hasZeroLength(x)
}
\arguments{
\item{x}{R object that has a length attribute, e.g. vector or list}
}
\value{
\code{TRUE} if \code{length(x) == 0L}, otherwise \code{FALSE}
}
\description{
This is just a "shortcut" to "length(x) == 0L"
}
\examples{
hasZeroLength(character()) # TRUE

hasZeroLength(list()) # TRUE

# Do not confuse, has nothing to do with the length of a string
# here: vector of length one
hasZeroLength("") # FALSE

# Remember that the length of a data frame is the number of its columns
hasZeroLength(data.frame(a = character())) # FALSE

}
