% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{headtail}
\alias{headtail}
\title{Print First and Last Rows of a Data Frame}
\usage{
headtail(x, n = 6, pattern = "[\%d rows omitted]")
}
\arguments{
\item{x}{data frame}

\item{n}{total number of rows to be printed.}

\item{pattern}{pattern given to \code{sprintf} containing a \code{\%d}
placeholder to print the number of omitted rows}
}
\value{
number of omitted rows, invisibly
}
\description{
Print the first and last rows of a data frame using head and tail, 
respectively. Print the number of omitted rows
}
\examples{
x <- data.frame(number = 1:26, letter = LETTERS)
headtail(x)
headtail(x, 10)
headtail(x, 16)
headtail(x[10:20, ], 10)

}
