% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDataFrame.R
\name{hsAddMissingCols}
\alias{hsAddMissingCols}
\title{Add missing columns to data frame}
\usage{
hsAddMissingCols(dataFrame, colNames, fill.value = NA)
}
\arguments{
\item{dataFrame}{data frame to which column names are to be appended}

\item{colNames}{vector containing names of columns that shall be contained in
\emph{dataFrame}}

\item{fill.value}{value to be inserted into newly created columns. Default: NA}
}
\value{
data frame with columns as listed in \emph{colNames}
}
\description{
Adds missing columns to the given data frame so that the resulting
  data frame contains all columns given in the vector \emph{colNames}.
  Added columns are filled with NA values.
}
