\name{hsAddToDict}
\alias{hsAddToDict}
\title{hsAddToDict}
\description{add assignements given in ... list to \emph{dictionary}}
\usage{hsAddToDict(dictionary = NULL, ...)}
\arguments{
  \item{dictionary}{
}
  \item{\dots}{
}
}



\author{Hauke Sonnenberg}





