\name{hsCountInStr}
\alias{hsCountInStr}
\title{hsCountInStr}
\description{Count occurrences of \emph{chr} in \emph{str}}
\usage{hsCountInStr(chr, str)}
\arguments{
  \item{chr}{
}
  \item{str}{
}
}

\value{number of orrurrences of \emph{char} in \emph{str}}

\author{Hauke Sonnenberg}





