\name{hsDelEmptyCols}
\alias{hsDelEmptyCols}
\title{Delete empty columns of data frame}
\description{Returns data frame in which all empty columns (NA in all rows) are removed}
\usage{hsDelEmptyCols(dataFrame)}
\arguments{
  \item{dataFrame}{data frame of which empty columns (NA in all rows) are to be removed}
}

\value{copy of input data frame but with all empty columns removed}

\author{Hauke Sonnenberg}





