% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hsMatrixToListForm}
\alias{hsMatrixToListForm}
\title{Convert "Matrix Form" (wide format) to "List Form" (long format)}
\usage{
hsMatrixToListForm(df, keyFields, parFields = setdiff(names(df),
  keyFields), colNamePar = "parName", colNameVal = "parVal",
  stringsAsFactors = FALSE)
}
\arguments{
\item{df}{data frame}

\item{keyFields}{names of key fields (e.g. date/time)}

\item{parFields}{names of fields representing differen parameters. Default:
column names that are not in \emph{keyFields}}

\item{colNamePar}{name of column in result data frame that will contain the
parameter names}

\item{colNameVal}{name of column in result data frame that will contain the
parameter values}

\item{stringsAsFactors}{if TRUE, columns of type character in the result data
frame are converted to factors. Parameter is passed to cbind, rbind.}
}
\value{
data frame in "list form" (long format)
}
\description{
Converts a data frame in "matrix form" to a data frame in "list form"
}
\seealso{
\code{\link[stats]{reshape}}
}
