\name{hsOpenWindowsExplorer}
\alias{hsOpenWindowsExplorer}
\title{open Windows Explorer}
\description{open Windows Explorer}
\usage{hsOpenWindowsExplorer(startdir = tempdir(), use.shell.exec = TRUE)}
\arguments{
  \item{startdir}{directory to be opened in Windows Explorer}
  \item{use.shell.exec}{
}
}



\author{Hauke Sonnenberg}





