% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsString.R
\name{hsQuoteChr}
\alias{hsQuoteChr}
\title{Quote Character Strings}
\usage{
hsQuoteChr(x, qchar = "'", escapeMethod = c("double", "backslash", "none"))
}
\arguments{
\item{x}{vector or list of character strings}

\item{qchar}{quoting character to be used. Default: single quote "'"}

\item{escapeMethod}{one of \code{"double", "backslash", "none"} deciding
how to treat the quote character if it occurs within the string to be
quoted}
}
\description{
quotes objects of type character with quoting character
}
