% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{hsRenameColumns}
\alias{hsRenameColumns}
\title{Rename Columns in a Data Frame (deprecated)}
\usage{
hsRenameColumns(dframe, renames)
}
\arguments{
\item{dframe}{data.frame}

\item{renames}{list with named elements each of which defines a column rename
in the form <old-name> = <new-name>}
}
\description{
Rename Columns in a Data Frame (deprecated, use renameColumns instead)
}
