\name{hsResolve}
\alias{hsResolve}
\title{hsResolve}
\description{Resolve strings according to substitutions defined in dictionary}
\usage{hsResolve(x, dict, dbg = FALSE)}
\arguments{
  \item{x}{(vector of) string expression(s) to be resolved using the dictionary
\emph{dict}.}
  \item{dict}{dictionary: list with named elements where the element name represents
the key and the element value represents the value assigned to the key.  }
  \item{dbg}{
}
}



\author{Hauke Sonnenberg}





