\name{hsRestoreAttributes}
\alias{hsRestoreAttributes}
\title{Restores object attributes}
\description{Restores given attributes that are not object attributes any more}
\usage{hsRestoreAttributes(x, attribs)}
\arguments{
  \item{x}{object}
  \item{attribs}{former attributes of x (as retrieved by attributes(x)) to be restored}
}



\author{Hauke Sonnenberg}





