% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsSystem.R
\name{hsShell}
\alias{hsShell}
\title{wrapper around "shell"}
\usage{
hsShell(commandLine, ...)
}
\arguments{
\item{commandLine}{character string passed to \code{shell}}

\item{...}{additional arguments passed to \code{shell}}
}
\description{
wrapper around "shell"
}
