% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsString.R
\name{hsSubstSpecChars}
\alias{hsSubstSpecChars}
\title{Substitution of Special Characters}
\usage{
hsSubstSpecChars(x, version = 1)
}
\arguments{
\item{x}{string containing special characters to be substituted}

\item{version}{integer number switching between different implementations.
version = 1 (default) uses double backslashes whether version = 2 uses
single backslashes in the replacement patterns.}
}
\value{
input string \emph{x} with special characters being substituted by 
  a meaningful represenation or underscore, multiple underscores replaced
  by a single underscore and multiple underscores at the end removed.
}
\description{
Substitution of special characters
}
