% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{hsSystem}
\alias{hsSystem}
\title{Wrapper around "system"}
\usage{
hsSystem(commandLine, ...)
}
\arguments{
\item{commandLine}{character string passed to \code{system}}

\item{...}{additional arguments passed to \code{system}}
}
\description{
Wrapper around "system"
}
