\name{hsTags}
\alias{hsTags}
\title{Find <tag>-tags in string}
\description{Return tags of the form <tag> that are contained in the string \emph{x}.}
\usage{hsTags(x, bt, dbg = FALSE)}
\arguments{
  \item{x}{
}
  \item{bt}{bracket type, must be one of c("<>", "[]")}
  \item{dbg}{
}
}



\author{Hauke Sonnenberg}





