\name{hsValidValue}
\alias{hsValidValue}
\title{hsValidValue}
\description{returns TRUE if text representation of number is in correct format in 
terms of decimal character and (optionally) thousand's separator character.}
\usage{hsValidValue(x, lng, dbg = FALSE, accept.na = TRUE)}
\arguments{
  \item{x}{
}
  \item{lng}{
}
  \item{dbg}{
}
  \item{accept.na}{
}
}



\author{Hauke Sonnenberg}





