% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indent.R
\name{indent}
\alias{indent}
\title{Indentat a String (Add Spaces to the Left)}
\usage{
indent(x, depth = 0L, tabLength = 2L)
}
\arguments{
\item{x}{vector of character}

\item{depth}{indentation level. Default: 0}

\item{tabLength}{number of spaces per indentation level. Default: 2}
}
\description{
Indentat a String (Add Spaces to the Left)
}
