\name{inRange}
\alias{inRange}
\title{inRange}
\description{check for values within minimum and maximum value}
\usage{inRange(values, min.value, max.value)}
\arguments{
  \item{values}{vector of values}
  \item{min.value}{minimum value (inclusive)}
  \item{max.value}{maximum value (inclusive)}
}

\value{vector of boolean}

\author{Hauke Sonnenberg}





