% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{intToNumeralSystem}
\alias{intToNumeralSystem}
\title{Convert Integers to Numeral System}
\usage{
intToNumeralSystem(x, base)
}
\arguments{
\item{x}{vector of integers}

\item{base}{base of the numeral system}
}
\value{
matrix with as many rows as there are elements in \code{x} and as 
  many columns as digits that are required to represent the integers in 
  \code{x} in the numeral system in base \code{base}. The elements of 
  \code{x} appear as row names whereas the powers of \code{base} appear as
  column names.
}
\description{
Convert Integers to Numeral System
}
\examples{
intToNumeralSystem(1:16, base = 2) # binary system
intToNumeralSystem(1:16, base = 10) # decimal system
intToNumeralSystem(1:16, base = 8) # octal system
}
