% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{isASCII}
\alias{isASCII}
\title{Do Strings Consist only of ASCII Characters?}
\usage{
isASCII(x)
}
\arguments{
\item{x}{vector of character}
}
\value{
vector of logical with \code{TRUE} at positions \code{i} where
\code{x[i]} contains only ASCII characters
}
\description{
The code has been "stolen" from \code{\link[tools]{showNonASCII}}
}
\examples{
months <- c("Januar", "Februar", "M\ue4rz")

cat(months, "\n")

isASCII(months)
}
