% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{isDotOrDoubleDot}
\alias{isDotOrDoubleDot}
\title{Does a string end with one or two dots (".", "..")?}
\usage{
isDotOrDoubleDot(x)
}
\arguments{
\item{x}{vector of character}
}
\value{
vector of logical
}
\description{
Does a string end with one or two dots (".", "..")?
}
\examples{

isDotOrDoubleDot(c("a", "b.", "c..", "d", "efg..h"))
}
