% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{isLoaded}
\alias{isLoaded}
\title{Has a script been loaded (with source)?}
\usage{
isLoaded(scriptName)
}
\arguments{
\item{scriptName}{name of the script for which to get the "loaded" state}
}
\description{
Check whether a script has been loaded (with source) or not. The information
  is read from a list that stores a logical value (\code{TRUE} or \code{FALSE}) for 
  each script name for which \code{\link{setLoaded}} has been called.
}
\examples{
  # For an example see kwb.utils::setLoaded()...
  
  # In fact, the information about loaded scripts is stored in the
  # R option "kwb.utils.loaded":
  setLoaded("myScript")
  
  options("kwb.utils.loaded")[[1]]
  

}
\seealso{
\code{\link{setLoaded}} (see example there)
}
