\name{isNaOrEmpty}
\alias{isNaOrEmpty}
\title{NA or the empty string ""?}
\description{is an object NA or equal to the empty string "" (after trimming)?}
\usage{isNaOrEmpty(x)}
\arguments{
  \item{x}{object to be tested for NA or being empty (equal to "", after trimming)}
}

\value{(vector of) logical, being TRUE for each element in \emph{x} that is
NA or the empty string "" (after trimming)}

\author{Hauke Sonnenberg}





