% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{isNetworkPath}
\alias{isNetworkPath}
\title{Does the Path Represent a Network Path?}
\usage{
isNetworkPath(x)
}
\arguments{
\item{x}{vector of character representing paths}
}
\description{
Does the Path Represent a Network Path?
}
\examples{
isNetworkPath("//server/folder/file.txt")

}
