\name{isNullOrEmpty}
\alias{isNullOrEmpty}
\title{isNullOrEmpty}
\description{isNullOrEmpty}
\usage{isNullOrEmpty(x)}
\arguments{
  \item{x}{object to be tested for NULL or being empty (vector or list of length 0 or 
data frame with no rows)}
}

\value{TRUE if x is NULL or x is a vector of length 0 or x is a data frame with
no rows.}

\author{Hauke Sonnenberg}





