% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isTryError.R
\name{isTryError}
\alias{isTryError}
\title{Does an Object Inherit from "try-error"?}
\usage{
isTryError(x)
}
\arguments{
\item{x}{R object}
}
\value{
logical of length one, TRUE if \code{x} inherits from "try-error",
  otherwise FALSE
}
\description{
Does an Object Inherit from "try-error"?
}
\examples{
result <- try(stop("Stop!"), silent = TRUE)
isTryError(result) # TRUE
}
