\name{kwb.utils-package}
\alias{kwb.utils-package}
\alias{kwb.utils}
\docType{package}
\title{some basic functions used by other kwb packages}
\description{some basic functions used by other kwb packages.}
\details{
\tabular{ll}{Package: \tab kwb.utils\cr
Version: \tab 0.2.1\cr
Title: \tab some basic functions used by other kwb packages\cr
Maintainer: \tab Hauke Sonnenberg <hauke.sonnenberg@kompetenz-wasser.de>\cr
Author: \tab Hauke Sonnenberg\cr
Imports: \tab PKI\cr
License: \tab GPL\cr}
}
\author{Hauke Sonnenberg}

\keyword{ package }


