% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{lastElement}
\alias{lastElement}
\title{Last Element}
\usage{
lastElement(x)
}
\arguments{
\item{x}{object}
}
\value{
The last element of \code{object} is returned: \code{x[[length(x)]]} 
  if \code{x} is a list (and not a data frame), otherwise \code{tail(x, 1)}.
}
\description{
Returns the last element using the function tail
}
