% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{left}
\alias{left}
\title{Left Part of a String}
\usage{
left(x, n)
}
\arguments{
\item{x}{vector of character}

\item{n}{number of characters to be kept from the beginning of each character
string within \code{x}}
}
\value{
vector of character
}
\description{
Left Part of a String
}
\examples{
left("Good Morning", 4)

}
