% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leftSubstringEquals.R
\name{leftSubstringEquals}
\alias{leftSubstringEquals}
\title{Is Left Substring of X Equal To Y?}
\usage{
leftSubstringEquals(x, y)
}
\arguments{
\item{x}{String of which the left part is compared with \code{y}}

\item{y}{String to be compared with the left part of \code{x}}
}
\description{
TODO: Simply use \code{\link{startsWith}}?
}
\examples{
leftSubstringEquals("Great job", "Great")
leftSubstringEquals("Great car", "Great")
leftSubstringEquals("Boring job", "Great")
}
