% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{limitToRange}
\alias{limitToRange}
\title{limit values to lie within [left, right]}
\usage{
limitToRange(x, left = .Machine$double.eps, right = 1)
}
\arguments{
\item{x}{vector of numeric values}

\item{left}{lower boundary}

\item{right}{upper boundary}
}
\description{
limit the values in \code{x} so that each value lies within the closed
interval \code{[left, right]}
}
\examples{
limitToRange(1:20, left = 5, right = 15)
}
