% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{listObjects}
\alias{listObjects}
\title{Get Names of Objects in .RData files}
\usage{
listObjects(files)
}
\arguments{
\item{files}{vector of full paths to .RData files}
}
\description{
Get Names of Objects in .RData files
}
\examples{
## Not run
  
## Search for available .RData files below "searchdir"
#searchdir <- "//poseidon/projekte$/SUW_Department/Projects/SEMA/WP/20_Braunschweig"
#files <- dir(searchdir, pattern = "\\\\.RData$", recursive = TRUE, full.names = TRUE)
  
## Get the names of the objects in the .RData files
#objectsInFiles <- listObjects(files)
  
## Which file contains the object "DataQ"?
#dataQ.found <- sapply(objectsInFiles, function(x) {"DataQ" \%in\% x})
  
#cat("DataQ was found in the following files:",
#    paste(files[dataQ.found], collapse = "\n  "))

}
