% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadFunctions.R
\name{loadFunctions}
\alias{loadFunctions}
\title{Load Functions from Scripts into Attached Namespace}
\usage{
loadFunctions(paths, attach = TRUE, name = "kwb", dbg = TRUE)
}
\arguments{
\item{paths}{vector of character with paths to files (R scripts) or 
directories (in which to look for R scripts).}

\item{attach}{if \code{TRUE} the environment with all parsed objects is
attached to the search path, otherwise the environment is returned}

\item{name}{name that is used when attaching the environment with all parsed
objects to the search path. Default: kwb}

\item{dbg}{logical indicating whether or not to show debug messages}
}
\value{
if \code{attach} is \code{FALSE} the environment with all parsed 
  objects is returned, otherwise \code{NULL}, invisibly.
}
\description{
This function parses R scripts at the given paths and provides all objects in
an environment that is attached to the search path.
}
