% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainClass.R
\name{mainClass}
\alias{mainClass}
\title{Main Class of an Object}
\usage{
mainClass(x)
}
\arguments{
\item{x}{any R object}
}
\value{
name of main class of \code{x} (vector of character of length one)
}
\description{
Returns the first element of what class(x) returns
}
\examples{
class(as.POSIXct("2022-06-02"))
mainClass(as.POSIXct("2022-06-02"))
}
