% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{matchesCriteria}
\alias{matchesCriteria}
\title{Do Data Frame Row Match Given Criteria?}
\usage{
matchesCriteria(
  Data,
  criteria = NULL,
  na.to.false = FALSE,
  add.details = FALSE,
  dbg = TRUE
)
}
\arguments{
\item{Data}{data frame}

\item{criteria}{vector of character containing conditions, in which the
column names of \code{Data}, e.g. \code{A} can appear unquoted, e.g.
\code{"A == 'x'"}}

\item{na.to.false}{if \code{TRUE} (the default is \code{FALSE}) NA in the
resulting vector will be replaced with FALSE}

\item{add.details}{if \code{TRUE} (the default is \code{FALSE}) a matrix 
containing the evaluation of each criterion is returned in attribute 
\code{details}}

\item{dbg}{if TRUE (default) for each criterion in \code{criteria} it is
shown for how many rows in \code{Data} the criterion is TRUE and for how
many rows it is FALSE}
}
\value{
vector of logical containing TRUE at positions representing rows in
  \code{Data} fulfilling the \code{conditions} and FALSE elsewhere
}
\description{
are data frame rows matching given criteria?
}
\examples{
# Define an example data frame
Data <- data.frame(A = c("x", "y", "z", NA),
                   B = c( NA,   2,   3, 4))

# Define one or more criteria
criteria <- c("A \%in\% c('y', 'z')", "B \%in\% 1:3")

# For which rows the criteria are met (vector of logical)?
matchesCriteria(Data, criteria, dbg = FALSE)

# You may use the function in the context of indexing:
Data[matchesCriteria(Data, criteria), ]

# Filtering for non-NA values
D1 <- Data[matchesCriteria(Data, "! is.na(A) & ! is.na(B)"), ]

# the same result is returned by:
D2 <- Data[matchesCriteria(Data, c("! is.na(A)", "! is.na(B)")), ]

identical(D1, D2)

}
