% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_data_frame.R
\name{matrixToDataFrame}
\alias{matrixToDataFrame}
\title{Convert a Matrix to a Data Frame (in "Long" Format)}
\usage{
matrixToDataFrame(
  x,
  name_row = NULL,
  name_column = NULL,
  name_value = "value",
  row_first = TRUE
)
}
\arguments{
\item{x}{matrix}

\item{name_row}{name to be given to the data frame column containing the
row "coordinates". Default: \code{names(dimnames(x))[1]} unless
\code{NULL}, \code{"row"} otherwise.}

\item{name_column}{name to be given to the data frame column containing the
column "coordinates". Default: \code{names(dimnames(x))[2]} unless
\code{NULL}, \code{"column"} otherwise.}

\item{name_value}{name to be given to the data frame column containing the
matrix values. Default: \code{value}}

\item{row_first}{if \code{TRUE} (the default), the "row column" will come
first, else the "column column".}
}
\value{
data frame with three columns: 1. row "coordinate", 2. column
  "coordinate", 3. value
}
\description{
Convert a Matrix to a Data Frame (in "Long" Format)
}
\examples{
m1 <- matrix(1:12, nrow = 3, dimnames = list(NULL, letters[1:4]))
m2 <- matrix(1:12, nrow = 3, dimnames = list(index = NULL, letters[1:4]))
m3 <- matrix(1:12, nrow = 3, dimnames = list(NULL, letter = letters[1:4]))

matrixToDataFrame(x = m1)
matrixToDataFrame(x = m1, row_first = FALSE)
matrixToDataFrame(x = m2)
matrixToDataFrame(x = m3)
matrixToDataFrame(x = m3, "myrow", "mycol", "myval")

}
