% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxStringLength.R
\name{maxStringLength}
\alias{maxStringLength}
\title{Maximum String Length in Vector of Character}
\usage{
maxStringLength(x)
}
\arguments{
\item{x}{vector of character}
}
\value{
integer representing the length of the longest string in \code{x}
}
\description{
Maximum String Length in Vector of Character
}
\examples{
maxStringLength(c("a", "ab", "abc", "x", "xy-z"))
}
