% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{mergeNamedArrays}
\alias{mergeNamedArrays}
\title{Merge List of Named Arrays}
\usage{
mergeNamedArrays(x, check_dim = TRUE)
}
\arguments{
\item{x}{list of arrays of the same dimension}

\item{check_dim}{logical. If \code{TRUE}, it is checked whether the source
dimension names are available in the target dimension names}
}
\description{
Merge List of Named Arrays
}
\examples{
a1 <- array(
  1:12,
  dim = c(2, 4, 2),
  dimnames = list(paste0("x", 1:2), paste0("y", 1:4), paste0("z", 1:2))
)

a2 <- array(
  11:16,
  dim = c(1, 3, 2),
  dimnames = list("x3", paste0("y", 2:4), paste0("z", 1:2))
)

mergeNamedArrays(list(a1, a2))

}
