% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{moveColumnsToFront}
\alias{moveColumnsToFront}
\title{Move Columns to the Left}
\usage{
moveColumnsToFront(x, columns = NULL)
}
\arguments{
\item{x}{data frame}

\item{columns}{vector of column names}
}
\value{
data frame or matrix with \code{columns} being the leftmost columns
}
\description{
Move columns of a data frame or matrix to the left
}
\examples{
x <- data.frame(a = 1:5, b = 2:6, c = 3:7)
  
moveColumnsToFront(x, "b")
moveColumnsToFront(x, c("b", "a"))
  
}
