% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{moveToFront}
\alias{moveToFront}
\title{Move elements to the start of a vector}
\usage{
moveToFront(x, elements)
}
\arguments{
\item{x}{vector}

\item{elements}{elements out of \code{x} to be moved to the front}
}
\value{
vector with \code{elements} coming first
}
\description{
Move elements to the start of a vector
}
\examples{
moveToFront(1:10, 5)
moveToFront(c("a", "b", "c", "x", "y", "d"), c("x", "y"))
  
}
