% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_readCsvInputFile.R
\name{msgAvailableFields}
\alias{msgAvailableFields}
\title{Message Listing Available Fields}
\usage{
msgAvailableFields(x)
}
\arguments{
\item{x}{vector of character}
}
\description{
Message to be shown if fields/columns are missing
}
