% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsSystem.R
\name{mySystemTime}
\alias{mySystemTime}
\title{Elapsed Time of a Function Call}
\usage{
mySystemTime(FUN, args)
}
\arguments{
\item{FUN}{function to be called}

\item{args}{list of arguments passed to \code{FUN}}
}
\description{
Call a function and show the elapsed time using \code{system.time})
}
