% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namedVectorFromColumns.R
\name{namedVectorFromColumns}
\alias{namedVectorFromColumns}
\title{Create Named Vector from two Columns of a Data Frame}
\usage{
namedVectorFromColumns(data, valueColumn, nameColumn)
}
\arguments{
\item{data}{a data frame}

\item{valueColumn}{name of column in \code{data} containing the names}

\item{nameColumn}{name of column in \code{data} containing the values}
}
\value{
vector with values in \code{data[[valueColumn]]}, named by
  \code{data[[nameColumn]]}
}
\description{
Create Named Vector from two Columns of a Data Frame
}
\examples{
data <- data.frame(a = 1:3, name = LETTERS[1:3])
namedVectorFromColumns(data, valueColumn = "a", nameColumn = "name")
}
