% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namedVectorToDataFrame.R
\name{namedVectorToDataFrame}
\alias{namedVectorToDataFrame}
\title{Convert a Named Vector to a Data Frame}
\usage{
namedVectorToDataFrame(x)
}
\arguments{
\item{x}{named vector}
}
\value{
data frame with columns \code{name}, containing the names of \code{x}
  and \code{value}, containing the values of \code{x}
}
\description{
Convert a Named Vector to a Data Frame
}
\examples{
namedVectorToDataFrame(c(a = 1, b = 2, c = 3))
}
