% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nNA.R
\name{nNA}
\alias{nNA}
\title{Number of NA values}
\usage{
nNA(x)
}
\arguments{
\item{x}{vector}
}
\value{
number of NA values in \code{x}
}
\description{
Number of NA values
}
\examples{
nNA(1:3)
nNA(c(1, NA, 3))
}
