% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{noFactorDataFrame}
\alias{noFactorDataFrame}
\title{Create Data Frame without Factors}
\usage{
noFactorDataFrame(...)
}
\arguments{
\item{\dots}{passed to \code{\link{data.frame}}}
}
\description{
Use \code{\link{data.frame}} with \code{stringsAsFactors = FALSE}
}
\examples{
# data.frame() by default converts character to factor
str(data.frame(id = 1:3, letter = LETTERS[1:3]))

# noFactorDataFrame keeps character as character
str(noFactorDataFrame(id = 1:3, letter = LETTERS[1:3]))

}
