% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noSuchElements.R
\name{noSuchElements}
\alias{noSuchElements}
\title{Message on Missing Elements}
\usage{
noSuchElements(x, available, type = "element", sorted = TRUE, suffix = "")
}
\arguments{
\item{x}{name of element that was not found}

\item{available}{names of elements that are available}

\item{type}{type of element to appear in the message. Default: "element"}

\item{sorted}{logical. Whether or not to print the available elements in
lexical order. Default: \code{TRUE}}

\item{suffix}{suffix to be appended to \code{type}. Can be used to
distinguish between singular and plural form. Default: \code{""}}
}
\description{
Message on Missing Elements
}
\examples{
cat(kwb.utils:::noSuchElements("x", LETTERS[1:3]))
cat(kwb.utils:::noSuchElements(c("x", "y"), LETTERS[1:3], suffix = "s"))
}
