% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nUnique.R
\name{nUnique}
\alias{nUnique}
\title{Number of Unique Values}
\usage{
nUnique(x)
}
\arguments{
\item{x}{R object that function \code{\link{unique}} can be applied on}
}
\value{
Number of unique values in \code{x} (integer)
}
\description{
Number of Unique Values
}
\examples{
nUnique(1:3)
nUnique(c(1, 1, 2))
}
