% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{objectSize}
\alias{objectSize}
\title{Object Size and Sizes of Sub Structures in Mb}
\usage{
objectSize(x, max_depth = 2, units = "auto", depth = 0)
}
\arguments{
\item{x}{object}

\item{max_depth}{number of sub levels of list structures to be shown}

\item{units}{passed to \code{\link{object.size}}}

\item{depth}{depth of recursive call (for internal use only)}
}
\value{
if \code{x} is a list, a list of the same structure is returned with
  each list elemenet replaced by its size, otherwise the object size of
  \code{x}. Each list or sub list with more than one element is assigned an
  attribute "total" to containing the total size of the list.
}
\description{
Object Size and Sizes of Sub Structures in Mb
}
