% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectToText.R
\name{objectToText}
\alias{objectToText}
\title{Convert R Object to Text Representation}
\usage{
objectToText(x)
}
\arguments{
\item{x}{R object}
}
\value{
vector of character representing the R code that reproduces the
  object \code{x}
}
\description{
Convert R Object to Text Representation
}
\examples{
objectToText(1:10)
objectToText((1:10)[-5])
cat(objectToText(head(iris)))
}
