% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orderDecreasinglyBy.R
\name{orderDecreasinglyBy}
\alias{orderDecreasinglyBy}
\title{Order Data Frame Decreasingly by one Column}
\usage{
orderDecreasinglyBy(df, column)
}
\arguments{
\item{df}{data frame}

\item{column}{name of column by which to order decreasingly.}
}
\description{
Order Data Frame Decreasingly by one Column
}
\examples{
(df <- data.frame(a = 1:3, b = 11:13))
orderDecreasinglyBy(df, "a")
}
