% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{pasteColumns}
\alias{pasteColumns}
\title{paste the columns of a data frame with a separator}
\usage{
pasteColumns(x, columns = names(x), sep = " ", ...)
}
\arguments{
\item{x}{data frame}

\item{columns}{names of columns to be pasted. Default: all columns}

\item{sep}{separator character. Default: space (" ")}

\item{\dots}{args passed to \code{\link{selectColumns}}, e.g. \code{do.stop} to control
whether the function shall stop if not all columns exist}
}
\value{
vector of character with each element representing the values of the
  selected columns of one row, being pasted with the separator character
}
\description{
paste the columns of a data frame with a separator
}
\examples{
  x <- data.frame(A = 1:3, B = 2:4)
  pasteColumns(x, sep = ";")
  

}
