% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{pasteColumns0}
\alias{pasteColumns0}
\title{Paste Columns of Data Frame without Separator}
\usage{
pasteColumns0(x, columns = names(x), ...)
}
\arguments{
\item{x}{data frame}

\item{columns}{names of columns to be pasted. Default: all columns}

\item{\dots}{args passed to \code{\link{pasteColumns}}}
}
\value{
vector of character with each element representing the values of the
  selected columns of one row, being pasted without a separator
}
\description{
Paste the columns of a data frame without a separator
}
\examples{
x <- data.frame(A = 1:3, B = 2:4)
pasteColumns0(x)

}
