% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{percentageOfSum}
\alias{percentageOfSum}
\title{Percentage of the Sum of Values}
\usage{
percentageOfSum(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector of numeric values}

\item{na.rm}{passed to \code{max}}
}
\value{
100 * x / sum(x)
}
\description{
Percentage of the Sum of Values
}
\examples{
p <- percentageOfSum(1:10)
stopifnot(sum(p) == 100)

}
