% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDataFrame.R
\name{posixColumnAtPosition}
\alias{posixColumnAtPosition}
\title{posixColumnAtPosition}
\usage{
posixColumnAtPosition(x)
}
\arguments{
\item{x}{data frame containing a date/time column}
}
\description{
posixColumnAtPosition
}
