% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{preparePdfIf}
\alias{preparePdfIf}
\title{Prepare PDF File if Condition is Met}
\usage{
preparePdfIf(to.pdf, PDF = "", ...)
}
\arguments{
\item{to.pdf}{condition determining whether \code{\link{preparePdf}} is
called or not}

\item{PDF}{full path to PDF file. If this is an empty string (default) a
temporary file is created and its path returned}

\item{\dots}{arguments passed to \code{\link{preparePdf}}}
}
\value{
full path to pdf file created if condition is met or "" else
}
\description{
Prepare PDF File if Condition is Met
}
