% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducibleSample.R
\name{print.repro_sample}
\alias{print.repro_sample}
\title{Print Method for Object of Class "repro_sample"}
\usage{
\method{print}{repro_sample}(x, ...)
}
\arguments{
\item{x}{object to be printed}

\item{\dots}{further arguments, not used.}
}
\description{
Print Method for Object of Class "repro_sample"
}
