% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode_decode.R
\name{printable_chars}
\alias{printable_chars}
\title{Different Sets of Printable ASCII Characters}
\usage{
printable_chars(level = 1)
}
\arguments{
\item{level}{one of 1, 2, or 3. Level 1 characters comprise the ten digits 0 
to 9 and 26 uppercase letters. Level 2 characters comprise the characters 
of level 1 as well as 26 lowercase letters. Level 3 characters comprise
altogether 88 printable characters.}
}
\value{
vector of character
}
\description{
Different Sets of Printable ASCII Characters
}
\examples{
printable_chars(1)
printable_chars(2)
printable_chars(3)

}
