\name{quotient}
\alias{quotient}
\title{quotient}
\description{calculate the quotient of two numbers}
\usage{quotient(dividend, divisor, substitute.value = Inf, warn = TRUE)}
\arguments{
  \item{dividend}{number to be devided}
  \item{divisor}{number by which dividend is to be devided}
  \item{substitute.value}{value to be returned if divisor is 0}
  \item{warn}{if TRUE, a warning is given if the divisor is zero}
}

\value{quotient of dividend and divisor: dividend/divisor}

\author{Hauke Sonnenberg}





