% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{randomValuesWithSum}
\alias{randomValuesWithSum}
\title{Vector of random Integer Values of given Sum}
\usage{
randomValuesWithSum(n, sumOfValues, names = seq_len(n))
}
\arguments{
\item{n}{number of values}

\item{sumOfValues}{sum of values in the result vector}

\item{names}{names of elements in the result vector. Default:
\code{seq_len(n)}}
}
\value{
named vector of integer values with \code{sum(values) == sumOfValues}
}
\description{
Vector of random Integer Values of given Sum
}
