% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rangeToSequence.R
\name{rangeToSequence}
\alias{rangeToSequence}
\title{Create Sequence from Range}
\usage{
rangeToSequence(x)
}
\arguments{
\item{x}{numeric vector with exactly two elements}
}
\value{
integer sequence between \code{x[1]} and \code{x[2]}
}
\description{
Create Sequence from Range
}
\examples{
rangeToSequence(c(1, 10))
}
