% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readArgs.R
\name{readArglists}
\alias{readArglists}
\title{Read Argument Lists from CSV File}
\usage{
readArglists(file = NULL,
  configTable = .readArglistsTable.csv(safePath(file), dbg = dbg),
  dbg = FALSE)
}
\arguments{
\item{file}{full path to comma separated file containing the argument specification}

\item{configTable}{Instead of the file you may provide a data frame containing the
configuration}

\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\description{
Read argument lists from CSV (or MS Excel) file
}
