% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{readDictionaries}
\alias{readDictionaries}
\title{Read Dictionary Files into a List of Dictionaries}
\usage{
readDictionaries(folder, pattern = "^dictionary_(.*)[.]txt$")
}
\arguments{
\item{folder}{path to the folder containing the files to be read}

\item{pattern}{regular expression to match the names of the files to be read.
The pattern is expected to contain a pair of parentheses around the part of
the file that shall be used as element name in the returned list}
}
\description{
Read files from a folder, specified by a file name pattern, into a list
of dictionaries
}
\seealso{
\code{\link{readDictionary}}
}
