\name{readDictionaryFromFile}
\alias{readDictionaryFromFile}
\title{readDictionaryFromFile}
\description{reads a dictionary (a list of "key = value"-pairs) from a text file.}
\usage{readDictionaryFromFile(dictionaryFile, sorted = TRUE)}
\arguments{
  \item{dictionaryFile}{full path to dictionary file}
  \item{sorted}{if TRUE (default) the entries in the dictionary will be sorted by their
keys}
}



\author{Hauke Sonnenberg}





