% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLinesWithEncoding.R
\name{readLinesWithEncoding}
\alias{readLinesWithEncoding}
\title{Read Lines by Giving the File Encoding}
\usage{
readLinesWithEncoding(file, ..., fileEncoding = "", encoding = "unknown")
}
\arguments{
\item{file}{a connection object or character string}

\item{\dots}{arguments passed to \code{\link{readLines}}}

\item{fileEncoding}{The name of the encoding to be assumed. Passed as
\code{encoding} to \code{\link{file}}, see there.}

\item{encoding}{passed to \code{\link{readLines}}.}
}
\description{
Read Lines by Giving the File Encoding
}
