% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readPackageFile}
\alias{readPackageFile}
\title{Read File from Package's extdata Folder}
\usage{
readPackageFile(file, package, stringsAsFactors = FALSE, ...)
}
\arguments{
\item{file}{name of file (without path)}

\item{package}{name of the package from which to read the file}

\item{stringsAsFactors}{passed to \code{\link[utils]{read.csv}} (default:
\code{FALSE})}

\item{\dots}{further arguments passed to \code{\link[utils]{read.csv}}}
}
\value{
result of reading \code{file} with \code{\link[utils]{read.csv}}
}
\description{
Read File from Package's extdata Folder
}
