% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{recursiveNames}
\alias{recursiveNames}
\title{names of all sublists of a list}
\usage{
recursiveNames(x, basename = "")
}
\arguments{
\item{x}{R list.}

\item{basename}{name to be used as prefix for all names found. Default: ""}
}
\description{
returns the names of all sublists of \emph{x} in the "$"-notation, e.g.
  list$sublist$subsublist$subsubsublist
}
