% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsMain.R
\name{recycle}
\alias{recycle}
\title{"recycle" vector to given length}
\usage{
recycle(x, n)
}
\arguments{
\item{x}{vector to be "recycled"}

\item{n}{target length}
}
\description{
recycle vector to given length
}
