% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{relativeCumulatedSum}
\alias{relativeCumulatedSum}
\title{relativeCumulatedSum}
\usage{
relativeCumulatedSum(values)
}
\arguments{
\item{values}{vector of numeric values}
}
\description{
relative cumulated sum of a vector of values
}
