\name{removeAttributes}
\alias{removeAttributes}
\title{Returns object without attributes}
\description{Returns object without attributes}
\usage{removeAttributes(x)}
\arguments{
  \item{x}{object}
}

\value{\emph{x}, but with its attributes removed}

\author{Hauke Sonnenberg}





