% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{removeDuplicates}
\alias{removeDuplicates}
\title{Remove Duplicated Values from a Vector}
\usage{
removeDuplicates(x, dbg = TRUE)
}
\arguments{
\item{x}{vector from which to remove duplicates}

\item{dbg}{if \code{TRUE} a debug message is shown}
}
\value{
\code{x} with duplicated values being removed
}
\description{
Remove Duplicated Values from a Vector
}
\examples{
removeDuplicates(c(1, 1, 2, 3, 4, 4))
}
