% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{removeElements}
\alias{removeElements}
\title{Remove Elements from a List}
\usage{
removeElements(x, elements)
}
\arguments{
\item{x}{a list}

\item{elements}{names of elements to remove}
}
\value{
\code{x} with elements with names given in \code{elements} being
  removed. User attributes of \emph{x} are restored.
}
\description{
Remove Elements from a List
}
