% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeEmpty.R
\name{removeEmpty}
\alias{removeEmpty}
\title{Remove Empty Elements from Vector}
\usage{
removeEmpty(x)
}
\arguments{
\item{x}{vector}
}
\value{
\code{x} with elements for which \code{\link{isNaOrEmpty}} is
  \code{TRUE} removed
}
\description{
Remove Empty Elements from Vector
}
\examples{
removeEmpty(c(1, NA, 3))
removeEmpty(c("a", "", "b", NA, "c", " ", "d"))
}
