% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeEmpty2.R
\name{removeEmpty2}
\alias{removeEmpty2}
\title{Remove Empty Elements from Vector of Character}
\usage{
removeEmpty2(x, dbg = FALSE)
}
\arguments{
\item{x}{vector of character}

\item{dbg}{if \code{TRUE} (the default is \code{FALSE}) a message about the 
number of removed elements is printed}
}
\description{
Remove Empty Elements from Vector of Character
}
\examples{
removeEmpty2(c("a", "", "c"))
removeEmpty2(c("a", "", "c", "", "e"), dbg = TRUE)
}
