% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{removeEmptyColumns}
\alias{removeEmptyColumns}
\title{Remove empty columns from a data frame}
\usage{
removeEmptyColumns(x, drop = FALSE, FUN = function(x) all(is.na(x)),
  dbg = TRUE)
}
\arguments{
\item{x}{data frame}

\item{drop}{if \code{TRUE} and only one column remains the column is returned
as a vector}

\item{FUN}{function called on each column to determine if all values in the
column are empty. Default: \code{function(x) all(is.na(x))}}

\item{dbg}{if \code{TRUE} debug messages are shown}
}
\value{
data frame \code{x} with empty columns (columns with NA in all rows) 
  being removed
}
\description{
Remove empty columns from a data frame.
}
\seealso{
\code{\link{hsDelEmptyCols}}
}
