% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{removeExtension}
\alias{removeExtension}
\title{Remove File Name Extension}
\usage{
removeExtension(x)
}
\arguments{
\item{x}{vector of character}
}
\description{
Remove File Name Extension
}
\examples{
removeExtension("example.R")
removeExtension("any/path/example.txt")

}
