% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeLeadingSlashes.R
\name{removeLeadingSlashes}
\alias{removeLeadingSlashes}
\title{Remove Leading Slashes of Strings}
\usage{
removeLeadingSlashes(x)
}
\arguments{
\item{x}{vector of character}
}
\value{
modified version of \code{x} with all leading slashes in all elements
  of \code{x} removed
}
\description{
Remove Leading Slashes of Strings
}
\examples{
removeLeadingSlashes(c("a", "/b", "//c"))
}
