% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsString.R
\name{removeSpaces}
\alias{removeSpaces}
\title{remove all spaces in string(s)}
\usage{
removeSpaces(x)
}
\arguments{
\item{x}{(vector of) character}
}
\value{
\emph{x} with all spaces removed
}
\description{
remove all spaces in string(s)
}
