% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{renameAndSelect}
\alias{renameAndSelect}
\title{rename and select columns of a data frame}
\usage{
renameAndSelect(data, renames, columns = unlist(renames))
}
\arguments{
\item{data}{data frame}

\item{renames}{list defining renames in the form of "oldName" = "newName" pairs}

\item{columns}{(new) names of colums to be selected}
}
\description{
rename and select columns of a data frame
}
