% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{repeated}
\alias{repeated}
\title{Repeated Substring}
\usage{
repeated(x, n)
}
\arguments{
\item{x}{substring to be repeated and pasted together to a new string}

\item{n}{number of times to repeat the substring}
}
\value{
vector of character of length one
}
\description{
Repeated Substring
}
\examples{
repeated("no ", 2)
repeated("yes ", 3)
repeated("yes no ", 3)

}
