% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceFileExtension.R
\name{replaceFileExtension}
\alias{replaceFileExtension}
\title{Replace File Name Extension}
\usage{
replaceFileExtension(file, extension)
}
\arguments{
\item{file}{file name or full path to a file}

\item{extension}{new file name extension (including the dot, e.g. ".txt")}
}
\description{
Replace the file name extension in a file name or file path
}
\examples{
replaceFileExtension(file = "run.bat", ".txt")
}
