% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reproducibleSample.R
\name{reproducibleSample}
\alias{reproducibleSample}
\title{Reproducible Call to the Sample Function}
\usage{
reproducibleSample(..., FUN = sample, random_seed = NULL)
}
\arguments{
\item{\dots}{arguments passed to \code{FUN}.}

\item{FUN}{the sample function to be called. Default: \code{\link{sample}}.}

\item{random_seed}{vector of integer as stored in \code{.Random.seed}.}
}
\value{
This function returns what \code{\link{sample}} returns with an 
  attribute \code{random_seed} attached.
}
\description{
This function calls \code{\link{sample}} and stores the values that are
required to reproduce the exact same sampling in the attribute 
\code{random_seed} of the returned vector. When this attribute is passed
to another call of this function, the values returned will be the same as
in the first call.
}
\examples{
# Take a sample
x <- reproducibleSample(1:100, 10)
x

# The full seed vector is returned in the attribute "random_seed"
random_seed <- attr(x, "random_seed")

# Take a new sample, this time passing the seed vector
y <- reproducibleSample(1:100, 10, random_seed = random_seed)
y

# The values are identical to the values of the first sampling
identical(x, y)

}
