% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDataFrame.R
\name{resetRowNames}
\alias{resetRowNames}
\title{Reset row names to 1:n}
\usage{
resetRowNames(x)
}
\arguments{
\item{x}{data frame or matrix}
}
\description{
Reset the row names of a data frame x to 1:nrow(x) by setting the
  \code{row.names} attribute to \code{NULL}.
}
\examples{
  persons <- data.frame(id = c(1, 2, 3), name = c("Peter", "Paul", "Mary"))
  
  persons.ordered <- persons[order(persons$name), ]
  
  # Original row names
  persons.ordered
  
  # Resetted row names
  resetRowNames(persons.ordered)
  

}
