% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{resolve}
\alias{resolve}
\title{Resolve string(s) using a dictionary}
\usage{
resolve(x, ...)
}
\arguments{
\item{x}{vector of character to be resolved or a list of which all elements
will be resolved using itself as a "dictionary". A dictionary is a list of 
\code{key = value} pairs defining string replacements.}

\item{\dots}{Unnamed arguments are treated as (further) dictionaries. These
are merged first to one dictionary before merging further (named) \code{key
= value} pairs.}
}
\description{
Resolve string(s) using a dictionary
}
\examples{
  
  file <- system.file("extdata", "dictionary.txt", package = "kwb.utils")
  
  dictionary <- readDictionary(file)
  
  # Resolve the dictionary
  resolve(dictionary)
  
  # Resolve the dictionary by setting an undefined placeholder
  resolve(dictionary, extension = "pdf")
    
  # Resolve a string
  resolve("dir.project", dictionary)
  
  # Set a placeholder "on-the-fly"
  resolve("file.out", dictionary, extension = "pdf")
  
  # Override a placeholder "on-the-fly"
  resolve("dir.project", dictionary, project = "new_project")
  
  # Resolve a vector of strings
  resolve(c("dir.root", "dir.project"), dictionary, project = "vector")

}
