% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{right}
\alias{right}
\title{Right Part of a String}
\usage{
right(x, n)
}
\arguments{
\item{x}{vector of character}

\item{n}{number of characters to be kept from the end of each character 
string within \code{x}}
}
\value{
vector of character
}
\description{
Right Part of a String
}
\examples{
right("Good Morning", 7)

}
