% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{rowOrColumnwisePercentage}
\alias{rowOrColumnwisePercentage}
\title{Rowwise or Columnwise Percentage}
\usage{
rowOrColumnwisePercentage(x, rowwise, default = 0, digits = 1)
}
\arguments{
\item{x}{two dimensional numeric data structure}

\item{rowwise}{if \code{TRUE} the percentage is calculated by row, else by
column}

\item{default}{default value to be used if the calculated percentage is 
\code{NA}.}

\item{digits}{number of digits (default: 1) to which the resulting 
percentages are to be rounded. Set to \code{NA} to suppress rounding}
}
\description{
Calculate the percentage (value divided by sum of values in the row/column) 
for each row/column
}
\seealso{
\code{\link{rowwisePercentage}}, \code{\link{columnwisePercentage}}
}
