% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{rowwisePercentage}
\alias{rowwisePercentage}
\title{Rowwise Percentage}
\usage{
rowwisePercentage(x, default = 0, digits = 1)
}
\arguments{
\item{x}{two dimensional numeric data structure}

\item{default}{default value to be used if the calculated percentage is 
\code{NA}.}

\item{digits}{number of digits (default: 1) to which the resulting 
percentages are to be rounded. Set to \code{NA} to suppress rounding}
}
\description{
Calculate the percentage (value divided by sum of values in the row) for 
each row
}
\examples{
# Create a random matrix of integer values
M1 <- matrix(sample(100, 12), nrow = 4, dimnames = list(LETTERS[1:4], 1:3))

# Introduce some NA
values <- as.numeric(M1)
values[sample(length(values), 3)] <- NA
M2 <- matrix(values, nrow = nrow(M1), dimnames = dimnames(M1))

M1
rowwisePercentage(M1)

M2
rowwisePercentage(M2)
rowwisePercentage(M2, default = 0)

}
