\name{rStylePath}
\alias{rStylePath}
\title{R compatible file path}
\description{R compatible file path with backslashes replaced with forward slashes}
\usage{rStylePath(path)}
\arguments{
  \item{path}{
}
}

\value{path in which backslashes are replaced with forward slashes}

\author{Hauke Sonnenberg}





