% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{runInDirectory}
\alias{runInDirectory}
\title{Change Working Directory and Run Function}
\usage{
runInDirectory(target.dir = tempdir(), FUN, ..., .dbg = FALSE)
}
\arguments{
\item{target.dir}{target directory. Default: \code{tempdir()}}

\item{FUN}{function to be invoked}

\item{\dots}{arguments to be passed to function FUN}

\item{.dbg}{if TRUE, debug messages on changing the directory are shown.
Default: \code{FALSE}}
}
\description{
Change working directory and run function
}
