% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDataFrame.R
\name{safeMerge}
\alias{safeMerge}
\title{Merge By Checking Column Existence}
\usage{
safeMerge(x, y, by = intersect(names(x), names(y)), by.x = by, by.y = by,
  ...)
}
\arguments{
\item{x}{just as in \code{\link[base]{merge}}}

\item{y}{just as in \code{\link[base]{merge}}}

\item{by}{just as in \code{\link[base]{merge}}}

\item{by.x}{just as in \code{\link[base]{merge}}}

\item{by.y}{just as in \code{\link[base]{merge}}}

\item{\dots}{additional arguments passed to \code{\link[base]{merge}}}
}
\description{
check existence of columns before merging
}
