% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{safePath}
\alias{safePath}
\title{Stop if Path does not Exist}
\usage{
safePath(...)
}
\arguments{
\item{\dots}{character vectors passed to \code{file.path}}
}
\value{
path as given in \code{path}
}
\description{
Check if file or folder path exists and return the path or stop if the path
does not exist
}
