\name{safeRowBind}
\alias{safeRowBind}
\title{"safe" rbind}
\description{rbind two data frames even if column names differ}
\usage{safeRowBind(dataFrame1, dataFrame2)}
\arguments{
  \item{dataFrame1}{
}
  \item{dataFrame2}{
}
}



\author{Hauke Sonnenberg}





