% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDataFrame.R
\name{safeRowBindAll}
\alias{safeRowBindAll}
\title{"safe" rbind of all data frames in a list}
\usage{
safeRowBindAll(x)
}
\arguments{
\item{x}{list of data frames}
}
\value{
data frame resulting from "rbind"-ing all data frames in \code{x}
}
\description{
rbind all data frames in a list using \code{\link{safeRowBind}}
}
