% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{safeRowBindOfListElements}
\alias{safeRowBindOfListElements}
\title{row-bind data frames in a list of lists}
\usage{
safeRowBindOfListElements(x, elementName)
}
\arguments{
\item{x}{list of lists each of which contains a data frame in element
\emph{elementName}}

\item{elementName}{name of list element in each sublist of \emph{x} which contains a
data frame}
}
\value{
data frame resulting from "row-binding" data frames.
}
\description{
row-bind data frames in a list of lists
}
\examples{
x <- list(
  list(
    number = 1,
    data = data.frame(x = 1:2, y = 2:3)
  ),
  list(
    number = 2,
    data = data.frame(x = 11:12, y = 12:13)
  )
)
  
safeRowBindOfListElements(x, "data")
  
# also working if the column names of the data frames in the "data" elements
# differ.
x[[1]]$data$z = 13:14
safeRowBindOfListElements(x, "data")

}
