% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{selectColumns}
\alias{selectColumns}
\title{Select Columns from a Data Frame}
\usage{
selectColumns(x, columns = NULL, drop = (length(columns) == 1),
  do.stop = TRUE)
}
\arguments{
\item{x}{data frame}

\item{columns}{vector of column names. If \code{columns} is of length 0 or
\code{NULL} (default) or \code{NA} \code{x} is returned unchanged.}

\item{drop}{if \code{TRUE} and if only one column is to be selected the
result is a vector (one dimensional) containing the values of the selected
column and not a data frame. One dimension has been \emph{dropped} then.
See the \code{help("[.data.frame")}. The default is \code{TRUE} if 
\code{length(columns) == 1}, else \code{FALSE}.}

\item{do.stop}{this flag controls whether the function stops (\code{do.stop =
TRUE}) or not (\code{do.stop = FALSE}) if there are non-existing columns to
be selected. If \code{do.stop = FALSE} only those columns are selected that
actually exist}
}
\value{
data frame containing the columns of \code{x} that are specified in 
  \code{columns} or \code{x} itself if \code{columns} is \code{NULL} or a
  vector containing the values of column \code{value} if \code{columns} is of
  length 1 and \code{drop = TRUE} (which is the default in this case).
}
\description{
Select columns from a data frame. Stop with message if columns do not exist
}
