% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsList.R
\name{selectElements}
\alias{selectElements}
\title{select (and rename) elements from list}
\usage{
selectElements(x, elements = NULL, do.stop = TRUE, do.warn = TRUE)
}
\arguments{
\item{x}{list}

\item{elements}{vector of element names. The names of named elements will be the names
in the output list}

\item{do.stop}{this flag controls whether the function stops (\code{do.stop = TRUE}) or
not (\code{do.stop = FALSE}) if there are non-existing elements to be
selected. If \code{do.stop = FALSE} only those elements are selected that
actually exist}

\item{do.warn}{if \code{TRUE} (default) and \code{do.stop = FALSE} a warning is given if
elements do not exist. Set to \code{FALSE} to suppress warnings}
}
\value{
list containing the elements of \code{x} that are specified in 
  \code{elements} or \code{x[[elements]]} if length of \code{elements} is 1 
  or \code{list()} if \code{elements} is empty. If the elements in vector 
  \code{elements} are named, these names are used in the output list.
}
\description{
select (and rename, if required) elements from list. Stop with message if
  elements do not exist
}
\examples{
  L <- list(a = 1, b = 2, c = 3, d = 4)
  
  # Select elements
  selectElements(L, c("a", "c"))
  
  # Select and rename at the same time
  selectElements(L, elements = c(a.new = "a", c.new = "c", "b"))
  

}
