% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{setMatrixColumns}
\alias{setMatrixColumns}
\title{set matrix column to value}
\usage{
setMatrixColumns(m, columnValuePairs, warn = TRUE)
}
\arguments{
\item{m}{matrix}

\item{columnValuePairs}{list of elements each of which defines an assignment in the form
<column-name> = <value>}

\item{warn}{if TRUE, warnings are given if columns named in \emph{columnValuePairs}
do not exist in matrix \emph{m}}
}
\description{
set matrix columns of given name to fix values
}
