% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{shorten}
\alias{shorten}
\title{Shorten Strings to a Maximum Length}
\usage{
shorten(x, max_chars = 10, delimiter = "...")
}
\arguments{
\item{x}{vector of character}

\item{max_chars}{maximum number of characters to which the strings in 
\code{x} are to be shortened}

\item{delimiter}{string to be used as separater between the start and the end 
of the strings in \code{x} that are longer than \code{max_chars} characters}
}
\value{
\code{x} with strings longer than \code{max_chars} characters being
  shortend by replacing characters in the centre by the \code{delimiter}
  string.
}
\description{
Shorten Strings to a Maximum Length
}
