% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showPackageObjects.R
\name{showPackageObjects}
\alias{showPackageObjects}
\title{Show Names of Objects/Functions in a Package}
\usage{
showPackageObjects(package = "kwb.utils", show = TRUE)
}
\arguments{
\item{package}{name of package of which to show the functions}

\item{show}{if \code{TRUE} (the default) the function names are printed to
the console otherwise the vector of function names is returned}
}
\description{
Show Names of Objects/Functions in a Package
}
