% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{sourceScripts}
\alias{sourceScripts}
\title{Load R Scripts with \code{source}}
\usage{
sourceScripts(scripts, dbg = TRUE)
}
\arguments{
\item{scripts}{full paths to R scripts}

\item{dbg}{if TRUE (default) log messages ("loading... ok") are shown}
}
\value{
the vector of script paths is returned invisibly
}
\description{
Load R Scripts with \code{source}
}
