% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{space}
\alias{space}
\title{Space String Used for Indentation}
\usage{
space(depth = 1L, tabLength = 2L)
}
\arguments{
\item{depth}{depth of indentation}

\item{tabLength}{number of spaces per indentation level}
}
\value{
vector of character of length one consisting of \code{depth *
  tabLength} space characters
}
\description{
Chain together \code{depth * tabLength} spaces
}
\examples{
cat(sprintf("\%s1\n\%s2\n\%s3\n", space(1), space(2), space(3)))
cat(sprintf("\%s1\n\%s2\n\%s3\n", space(1, 4), space(2, 4), space(3, 4)))

}
