% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array.R
\name{splitAlongDim}
\alias{splitAlongDim}
\title{Split Array Along a Dimension}
\usage{
splitAlongDim(a, n)
}
\arguments{
\item{a}{an array}

\item{n}{number of the dimension along which to split the array}
}
\value{
array of one dimension less than \code{a}

# Define an array
A <- array(1:8, dim = c(2, 2, 2), dimnames = list(
  paste0("x", 1:2), paste0("y", 1:2), paste0("z", 1:2)
))

splitAlongDim(A, 1)
splitAlongDim(A, 2)
splitAlongDim(A, 3)
}
\description{
Split an array along its n-th dimension. The implementation was found here:
https://stackoverflow.com/questions/20198751/three-dimensional-array-to-list
}
