% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{splitIntoFixSizedBlocks}
\alias{splitIntoFixSizedBlocks}
\title{split into blocks of same size}
\usage{
splitIntoFixSizedBlocks(data, blocksize)
}
\arguments{
\item{data}{data frame or matrix}

\item{blocksize}{number of rows in each block into which \code{data} is split}
}
\value{
list of data frames (if \code{data} is a data frame) or list of matrices
  (if \code{data} is a matrix)
}
\description{
split a data frame or matrix into blocks of the same size (= data frames of
  matrices with the same number ofrows)
}
