% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockRange.R
\name{startsToRanges}
\alias{startsToRanges}
\title{Row Numbers of Start Rows to From/To Row Ranges}
\usage{
startsToRanges(starts, lastStop, startOffset = 1, stopOffset = 1)
}
\arguments{
\item{starts}{integer vector of start indices}

\item{lastStop}{integer value of the last stop index}

\item{startOffset}{integer offset applied to the starts}

\item{stopOffset}{integer offsets applied to the ends}
}
\value{
data frame with columns \code{from} and \code{to}
}
\description{
A vector of row numbers is transformed to a data frame describing row ranges 
by numbers of  first and last rows
}
\examples{
starts <- c(1, 10, 20, 35)

ok <- identical(
  startsToRanges(starts, lastStop = 50), 
  data.frame(
    from = c(2, 11, 21, 36),
    to = c(9, 19, 34, 50)
  )
)
  
ok <- ok && identical(
  startsToRanges(starts, lastStop = 55, startOffset = 2, stopOffset = 2), 
  data.frame(
    from = c(3, 12, 22, 37),
    to = c(8, 18, 33, 55)
  )
)
  
ok
  
}
