% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopFormatted.R
\name{stopFormatted}
\alias{stopFormatted}
\title{Stop Function Exectution With Formatted Message}
\usage{
stopFormatted(x, ..., call. = FALSE)
}
\arguments{
\item{x}{Error message, possibly containing percentage placeholders, passed
as argument \code{fmt} to \code{\link{sprintf}}}

\item{\dots}{as many values as referenced with percentage placeholders in 
the error message}

\item{call.}{logical, indicating if the call should become part of the
    error message.}
}
\description{
Stop Function Exectution With Formatted Message
}
\examples{
try(stopFormatted(
  "Hi, \%s, the program fails for the \%d-th time.", "Dexter", 1000
))
}
