% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{stopIfNotMatrix}
\alias{stopIfNotMatrix}
\title{Stop with a Message if Input is not a Matrix}
\usage{
stopIfNotMatrix(x)
}
\arguments{
\item{x}{object to be checked with \code{\link{is.matrix}}}
}
\description{
Stop with a Message if Input is not a Matrix
}
