\name{stringContains}
\alias{stringContains}
\title{stringContains}
\description{stringContains}
\usage{stringContains(x, contains)}
\arguments{
  \item{x}{
}
  \item{contains}{
}
}



\author{Hauke Sonnenberg}




\examples{
stringContains(c("abc", "Kabeljau", "Arabella"), "ab")
stringContains(c("abc", "Kabeljau", "Arabella"), "abc")
}
