\name{stringEndsWith}
\alias{stringEndsWith}
\title{stringEndsWith}
\description{stringEndsWith}
\usage{stringEndsWith(x, endsWith)}
\arguments{
  \item{x}{
}
  \item{endsWith}{string to be searched for at the end of the string(s) in \emph{x}  }
}



\author{Hauke Sonnenberg}




\examples{
stringEndsWith(c("abc", "Kabeljau", "Arabella"), "a")
stringEndsWith(c("abc", "Kabeljau", "Arabella"), "jau")
}
