% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{stringList}
\alias{stringList}
\title{string of comma separated quoted strings}
\usage{
stringList(x, qchar = "'", collapse = ", ")
}
\arguments{
\item{x}{vector of character}

\item{qchar}{character to be used for quoting, default: single quote character}

\item{collapse}{characters used to separate the strings. Default: ", "}
}
\description{
create a string of comma separated quoted strings
}
