\name{stringStartsWith}
\alias{stringStartsWith}
\title{stringStartsWith}
\description{stringStartsWith}
\usage{stringStartsWith(x, startsWith)}
\arguments{
  \item{x}{
}
  \item{startsWith}{string to be searched for at the beginning of the string(s) in \emph{x}}
}



\author{Hauke Sonnenberg}




\examples{
stringStartsWith(c("abc", "Kabeljau", "Arabella"), "ab")
stringStartsWith(c("abc", "Kabeljau", "Arabella"), "A")
}
