% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsString.R
\name{stringToExpression}
\alias{stringToExpression}
\title{Convert String to Expression}
\usage{
stringToExpression(expressionString)
}
\arguments{
\item{expressionString}{character string to be converted to an expression}
}
\description{
Convert String to Expression
}
