% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsLookup.R
\name{tableLookup}
\alias{tableLookup}
\title{lookup value for key in table}
\usage{
tableLookup(x, key, default = NULL)
}
\arguments{
\item{x}{data frame with at least two columns. Keys are expected to be in the first
and values are expected to be in the second column, respectively.}

\item{key}{key for which a value is to be looked up}

\item{default}{default value that is returned if the key is not a key of the lookup table}
}
\value{
value looked up in the lookup table of default if \code{key} is not 
  contained in the first column of the lookup table \code{x}
}
\description{
lookup a value in the second column of a data frame \code{x} where the value
  in the first column matches the \code{key}.
}
