% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textToObject.R
\name{textToObject}
\alias{textToObject}
\title{Convert Text Representation Back to R Object}
\usage{
textToObject(x)
}
\arguments{
\item{x}{vector of character as returned by \code{\link{objectToText}}}
}
\value{
R object
}
\description{
Convert Text Representation Back to R Object
}
\examples{
textToObject("c(1:10)")
stopifnot(identical(iris, textToObject(objectToText(iris))))
}
