% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toConditional.R
\name{toConditional}
\alias{toConditional}
\title{Convert Function to Function that is Called if Condition is Met}
\usage{
toConditional(FUN)
}
\arguments{
\item{FUN}{a function}
}
\description{
Convert Function to Function that is Called if Condition is Met
}
\examples{
square <- function(x) x^2
negate <- function(x) -x
`\%>\%` <- magrittr::`\%>\%`
do_square <- TRUE
do_negate <- TRUE
10 \%>\%
  toConditional(square)(do_square) \%>\%
  toConditional(negate)(do_negate)
}
