% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{toFactor}
\alias{toFactor}
\title{Convert to Factor with unique Values as Levels}
\usage{
toFactor(x)
}
\arguments{
\item{x}{vector to be converted to factor}
}
\description{
In contrast to the default behaviour of \code{\link[base]{as.factor}}, this 
function uses the unsorted unique values of \code{x} as levels and not sorted
unique values.
}
\examples{
x <- c("b", "c", "a")

as.factor(x) # Levels: a b c

toFactor(x) # Levels: b c a
}
