% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsConvert.R
\name{toFormula}
\alias{toFormula}
\title{create formula from left and right term strings}
\usage{
toFormula(leftterm, rightterms, as.formula = TRUE)
}
\arguments{
\item{leftterm}{character. Left term of the formula}

\item{rightterms}{vector of character. Right terms of the formula to be concatenated with
"+".}

\item{as.formula}{if TRUE (default) the formula object is returned, otherwise the formula
string (character)}
}
\value{
formula object as generated by \code{formula} or formula string 
  (\code{character}) of the form \code{leftterm ~ rightterms[1] +
  rightterms[2] + ...} if \code{as.formula} is FALSE
}
\description{
create a formula of the form \code{leftterm ~ rightterms[1] + rightterms[2]
  + ...}
}
\examples{
  f1 <- toFormula("y", c("x1", "x2"))
  f2 <- toFormula("y", paste0("x", 1:20))
  
  f3 <- toFormula("BMI", c("height", "mass"), as.formula = FALSE)
  
  # f1 and f2 are formulas ...
  class(f1)
  class(f2)
  
  # ... but f3 is just "character"
  class(f3)
  

}
