% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{toInches}
\alias{toInches}
\title{Convert Centimeters to Inches}
\usage{
toInches(cm)
}
\arguments{
\item{cm}{vector of numeric representing length(s) in cm}
}
\value{
vector of numeric representing length(s) in inches
}
\description{
Convert Centimeters to Inches
}
