% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{toLookupList}
\alias{toLookupList}
\title{Keys and Values to Lookup List}
\usage{
toLookupList(keys, values, data = NULL)
}
\arguments{
\item{keys}{vector of character representing the keys of the dictionary}

\item{values}{vector of character representing the values of the dictionary}

\item{data}{Optional. Data frame with two columns of which the first is
assumed to contain the keys and the second is assumed to contain the values
of the dictionary}
}
\value{
dictionary: list with \emph{values} as elements and \emph{keys} as
  names
}
\description{
Keys and values to lookup list (list with elements representing the values 
and the names of the list elements representing the corresponding keys)
}
