% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{toPositiveIndices}
\alias{toPositiveIndices}
\title{Negative Indices to Positive Indices}
\usage{
toPositiveIndices(indices, n)
}
\arguments{
\item{indices}{vector of integer indices}

\item{n}{maximum index. The index -1 is mapped to \code{n}, the index -2 to
\code{n - 1}, etc.}
}
\description{
Convert negative indices to positive indices where negative indices are 
interpreted as indices counting from the back of the vector
}
