% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{underscoreToPercent}
\alias{underscoreToPercent}
\title{Replace underscore with percent sign}
\usage{
underscoreToPercent(x)
}
\arguments{
\item{x}{character vector containing underscores}
}
\description{
Replace underscore with percent sign. May be used to define time format
  strings as defaults in function declarations which are not supported
  by inlinedocs.
}
