% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniqueDirnames.R
\name{uniqueDirnames}
\alias{uniqueDirnames}
\title{Unique Directory Names}
\usage{
uniqueDirnames(x)
}
\arguments{
\item{x}{vector of file paths}
}
\value{
vector of character with the unique directories found in \code{x}
}
\description{
Unique Directory Names
}
\examples{
uniqueDirnames(c("a/b", "a/c", "b/c", "b/d"))
}
