% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsSystem.R
\name{user}
\alias{user}
\title{name of the current user}
\usage{
user()
}
\value{
character string represenging the name of the current user
}
\description{
Get the name of the current user from the environment variables
}
