\name{warnIfEmpty}
\alias{warnIfEmpty}
\title{warnIfEmpty}
\description{Gives a warning if the object is NULL or empty and returns the object}
\usage{warnIfEmpty(x)}
\arguments{
  \item{x}{object to be tested for NULL or being empty (vector of length 0 or 
data frame with no rows)}
}



\author{Hauke Sonnenberg}





