% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{warningDeprecated}
\alias{warningDeprecated}
\title{Create Warning About a Deprecated Function}
\usage{
warningDeprecated(old_name, new_name)
}
\arguments{
\item{old_name}{name of deprecated function}

\item{new_name}{name of new function to be used instead}
}
\description{
Create Warning About a Deprecated Function
}
\examples{
warningDeprecated("old_function()", "new_function()")

}
