% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichExecutables.R
\name{whichExecutables}
\alias{whichExecutables}
\title{Find Executable Files in Directories given in PATH Environment Variable}
\usage{
whichExecutables(pattern = NULL, full.names = FALSE)
}
\arguments{
\item{pattern}{if not NULL, the result list is filtered for file names 
matching the pattern}

\item{full.names}{if TRUE, the list items are the full paths, otherwise the
file names only}
}
\value{
list with one entry per directory from the PATH variable
}
\description{
Find Executable Files in Directories given in PATH Environment Variable
}
