\name{windowsPath}
\alias{windowsPath}
\title{convert to MS Windows-compatible path}
\description{create MS Windows-compatible path by substituting forward slashes with 
backslashes}
\usage{windowsPath(path)}
\arguments{
  \item{path}{
}
}



\author{Hauke Sonnenberg}





