% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{writeDictionary}
\alias{writeDictionary}
\title{Write a Dictionary (List) to a Text File}
\usage{
writeDictionary(dictionary, file)
}
\arguments{
\item{dictionary}{list of character vectors of length one defining \code{key
= value} pairs forming a dictionary.}

\item{file}{full path to the file to which \code{dictionary} is to be written}
}
\description{
Write a Dictionary (List) to a Text File
}
\seealso{
\code{\link{readDictionary}, \link{readDictionaries}}
}
