% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{writeText}
\alias{writeText}
\title{Write Text Lines to a File}
\usage{
writeText(x, file, type = "", dbg = TRUE, ...)
}
\arguments{
\item{x}{vector of character representing the lines to be written to
\code{file}, passed to \code{\link{writeLines}}}

\item{file}{path to file to be written, passed to \code{\link{writeLines}}}

\item{type}{character string to be included in the debug message:
"Writing <type>'file-path' ..."}

\item{dbg}{if \code{TRUE}, debug messages are shown}

\item{\dots}{further arguments passed to \code{\link{writeLines}}}
}
\description{
Write text to a file using \code{\link{writeLines}} and output a debug
message by default
}
\examples{
# Set path to a temporary file
file <- tempfile()

# Define text to be written to file
x <- c("Hello", "world")

# Write text to the the temporary file
writeText(x, file)

# Make the debug message more informative
writeText(x, file, type = "welcome file")

# Read lines back and show on the console
catLines(readLines(file))

}
