% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{addRowWithName}
\alias{addRowWithName}
\title{Add a Row with a Name}
\usage{
addRowWithName(x, y, row.name)
}
\arguments{
\item{x}{data frame to which row is to be appended}

\item{y}{data frame containing the row to be appended (exacly one row expected)}

\item{row.name}{name of row to be given in result data frame}
}
\value{
\emph{x} with row of \emph{y} (named \emph{row.name}) appended to it
}
\description{
add row to data frame and give a row name at the same time
}
