% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{atLeastOneRowIn}
\alias{atLeastOneRowIn}
\title{at least one row in data frame}
\usage{
atLeastOneRowIn(dframe)
}
\arguments{
\item{dframe}{data frame}
}
\description{
returns TRUE if data frame has at least one row, else FALSE
}
