% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{catIf}
\alias{catIf}
\title{call cat if condition is met}
\usage{
catIf(condition, ...)
}
\arguments{
\item{condition}{if TRUE, cat is called, else not}

\item{\dots}{arguments passed to cat}
}
\description{
call cat if condition is met
}
