% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsIO.R
\name{catLines}
\alias{catLines}
\title{print character vector to the console}
\usage{
catLines(x)
}
\arguments{
\item{x}{vector of character representing text lines to be printed}
}
\description{
call cat on character vector, pasted with collapse = <new line>
}
