\name{colStatisticOneFunction}
\alias{colStatisticOneFunction}
\title{colStatisticOneFunction}
\description{applies a statistical function to all columns of a data frame}
\usage{colStatisticOneFunction(dataFrame, FUN, na.rm = FALSE)}
\arguments{
  \item{dataFrame}{
}
  \item{FUN}{statistical function to be applied on each column of dataFrame
possible values: "sum", "mean", "min", "max", "number.na" (number of NA 
values), "length" (number of values)  }
  \item{na.rm}{if TRUE, NA values are removed before applying the statistical function  }
}



\author{Hauke Sonnenberg}





