\name{columnDescriptor}
\alias{columnDescriptor}
\title{columnDescriptor}
\description{columnDescriptor}
\usage{columnDescriptor(match = ".*", fixed = FALSE)}
\arguments{
  \item{match}{pattern or fixed text to match in header line}
  \item{fixed}{if TRUE, \emph{match} is taken as a fixed string to be looked for in the
header line, otherwise it is interpreded as a regular expression}
}



\author{Hauke Sonnenberg}





