% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{compareDataFrames}
\alias{compareDataFrames}
\title{Compare two data frames by columns}
\usage{
compareDataFrames(x, y)
}
\arguments{
\item{x}{first data frame}

\item{y}{second data frame}
}
\value{
list of logical
}
\description{
Compare two data frames by columns
}
\examples{
  x <- data.frame(a = 1:2, b = 2:3)
  y <- x
  
  test1 <- all(unlist(compareDataFrames(x, y)))
  
  z <- compareDataFrames(x, y[, c("b", "a")])
  expectedFalse <- c("identical", "identicalExceptAttributes", "sameColumnNames")
  test2 <- all(names(which(!unlist(z))) == expectedFalse)
  
  test1 && test2
  
}
