% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsSystem.R
\name{createDirAndReturnPath}
\alias{createDirAndReturnPath}
\title{create directory if it does not exist}
\usage{
createDirAndReturnPath(dir.to.create, dbg = TRUE, confirm = FALSE)
}
\arguments{
\item{dir.to.create}{character string representing the path to the directory
to be created}

\item{dbg}{if \code{TRUE} messages about created or found directories are
shown}

\item{confirm}{if \code{TRUE} (the default is \code{FALSE}!) the user is
asked to confirm the creation of a directory}
}
\description{
create directory if it does not exist
}
