% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{createMatrix}
\alias{createMatrix}
\title{Matrix with Row and Column Names}
\usage{
createMatrix(rowNames, colNames = rowNames, value = 0,
  name.row = NULL, name.col = NULL)
}
\arguments{
\item{rowNames}{character vector of row names to be given to the matrix}

\item{colNames}{character vector of column names to be given to the matrix}

\item{value}{value to be given to each matrix element}

\item{name.row}{optional. Name to be given to the row dimension}

\item{name.col}{optional. Name to be given to the column dimension}
}
\value{
matrix with \code{rowNames} as row names and \code{colNames} as
  column names, filled with \emph{value} at each position
}
\description{
Create a matrix by giving row and column names and with all elements being 
set to a default value
}
\examples{
## Initialise a matrix with rows A to E and columns x to z of value -1
createMatrix(c("A", "B", "C", "D", "E"), c("x", "y", "z"), -1)

## By default the column names are assumed to be equal to the row names
createMatrix(c("A", "B", "C"))

## Initialise a square matrix with NA
createMatrix(c("A", "B", "C"), value = NA)

## Give a name to the row dimension
createMatrix(c("A", "B", "C"), name.row = "Letters")

}
