% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{createPasswordFile}
\alias{createPasswordFile}
\title{Create a File Storing a Password in Encrypted Form}
\usage{
createPasswordFile(account = NULL, keyFile, passwordFile, password = NULL)
}
\arguments{
\item{account}{optional. Name of account the user is asked to enter the 
password for. Only used for the input prompt.}

\item{keyFile}{path to the file containing the encryption/decryption key,
as generated by \code{\link{generateKeyFile}}}

\item{passwordFile}{path to the password file to be created. An existing file
will be overwritten!}

\item{password}{password for account. If \code{NULL} (default) the user will
be asked to enter the password on the console}
}
\description{
Create a File Storing a Password in Encrypted Form
}
\seealso{
\code{\link{generateKeyFile}}, \code{\link{getPassword}}
}
