% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsEncryption.R
\name{createPasswordFile}
\alias{createPasswordFile}
\title{create encrypted password file for account}
\usage{
createPasswordFile(account, keyFile, passwordFile)
}
\arguments{
\item{account}{name of account the user is asked to enter the password for}

\item{keyFile}{path to the file containing the encryption/decryption key}

\item{passwordFile}{path to the password file}
}
\description{
create encrypted password file for account
}
