\name{createPasswordFile}
\alias{createPasswordFile}
\title{create encrypted password file for account}
\description{create encrypted password file for account}
\usage{createPasswordFile(account, keyFile, passwordFile)}
\arguments{
  \item{account}{name of account the user is asked to enter the password for}
  \item{keyFile}{file containing the encryption/decryption key}
  \item{passwordFile}{
}
}



\author{Hauke Sonnenberg}





