% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStorage.R
\name{createStorage}
\alias{createStorage}
\title{Create Storage Object to Save/Load RDS files}
\usage{
createStorage(path)
}
\arguments{
\item{path}{path to the folder in which to store the RDS files. If the folder 
does not exist it is attempted to be created.}
}
\value{
list with "member functions" \code{list()}, \code{save()}, 
  \code{load()}, \code{remove()}
}
\description{
This function returns an object that implements the methods \code{list()}, 
\code{save()}, \code{load()}, \code{remove()} that list, store, load or 
remove, respectively RDS files. The object is associated to a folder that 
is given to this function.
}
\examples{
# Create a storage object
storage <- createStorage(tempdir())

# At the beginning, the storage is empty
storage$list()

# Store objects
a <- 2
b <- "two"
storage$save(a, b, c = list("Hello", "World"))

# List the objects
storage$list()

# Retrieve an object
storage$load("c")

# Remove objects
storage$remove("a")
storage$remove("b")
storage$remove("c")

# The storage is empty again
storage$list()

}
