% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{csvTextToDataFrame}
\alias{csvTextToDataFrame}
\title{CSV Text to Data Frame}
\usage{
csvTextToDataFrame(text, ...)
}
\arguments{
\item{text}{character vector representing lines of comma separated values}

\item{\dots}{arguments passed to \code{\link[utils]{read.table}}}
}
\description{
CSV Text to Data Frame
}
