% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsString.R
\name{csvTextToDataFrame}
\alias{csvTextToDataFrame}
\title{csvTextToDataFrame}
\usage{
csvTextToDataFrame(text, ...)
}
\arguments{
\item{text}{character vector representing lines of comma separated values}

\item{\dots}{arguments passed to \code{utils::read.table}}
}
\description{
csvTextToDataFrame
}
