% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{enlargeVector}
\alias{enlargeVector}
\title{Enlarge a Vector to given Length}
\usage{
enlargeVector(x, length.out, fill.with = "")
}
\arguments{
\item{x}{vector}

\item{length.out}{desired length of output vector}

\item{fill.with}{element to fill the vector up with (default: "")}
}
\value{
\code{x}, filled up with \code{fill.with} to a final length of
  \code{length.out}
}
\description{
Enlarge a vector to the given length, filling with given element
}
\examples{
kwb.utils::enlargeVector(1:5, 10, fill.with = 0) 
kwb.utils::enlargeVector(1:5, 10, fill.with = NA) 
kwb.utils::enlargeVector(c("a", "b", "c"), 10) 
kwb.utils::enlargeVector(c("a", "b", "c"), 10, fill.with = "?")

}
