% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsList.R
\name{excludeNULL}
\alias{excludeNULL}
\title{Exclude all NULL entries from a list}
\usage{
excludeNULL(x)
}
\arguments{
\item{x}{a list}
}
\value{
list \code{x} with all \code{NULL} entries excluded
}
\description{
Exclude all null entries from a list
}
\examples{
  L <- list(a = 1, b = NULL, c = "three")
  L
  
  excludeNULL(L)
  

}
