% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createFunctionExtdataFile.R
\name{extdataFile}
\alias{extdataFile}
\title{Path to File in Installed Package}
\usage{
extdataFile(...)
}
\arguments{
\item{\dots}{parts of the file path to be passed to \code{\link{system.file}}}
}
\value{
path to file in the package installation folder in the R library
  or "" if the path does not exist
}
\description{
Path to File in Installed Package
}
\examples{
# List the files provided in the "extdata" folder of kwb.utils
dir(extdataFile())
}
