% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{finishAndShowPdfIf}
\alias{finishAndShowPdfIf}
\title{Finish and Display PDF File if Condition is Met}
\usage{
finishAndShowPdfIf(to.pdf, PDF, ...)
}
\arguments{
\item{to.pdf}{if \code{TRUE} the pdf device is closed and the pdf file is
opened in a viewer}

\item{PDF}{path to the opened pdf file}

\item{...}{arguments passed to \code{\link{finishAndShowPdf}}}
}
\description{
Finish and Display PDF File if Condition is Met
}
