% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsMain.R
\name{firstElement}
\alias{firstElement}
\title{first element}
\usage{
firstElement(x)
}
\arguments{
\item{x}{object}
}
\value{
first element: x[1]
}
\description{
Returns the first element using the function head
}
