% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsConvert.R
\name{frenchToAscii}
\alias{frenchToAscii}
\title{French unicode letter to ASCII letter(s)}
\usage{
frenchToAscii()
}
\value{
list of ASCII characters (list elements) replacing unicode characters
  (element names)
}
\description{
French unicode letter to ASCII letter(s)
}
