% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{frenchToAscii}
\alias{frenchToAscii}
\title{French Unicode Letter to ASCII Letter(s)}
\usage{
frenchToAscii()
}
\value{
list of ASCII characters (list elements) replacing unicode characters
  (element names)
}
\description{
French Unicode Letter to ASCII Letter(s)
}
