% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsEncryption.R
\name{generateKeyFile}
\alias{generateKeyFile}
\title{generate a decryption key file}
\usage{
generateKeyFile(target)
}
\arguments{
\item{target}{full path to the file to which the key shall be written}
}
\description{
generate a decryption key file
}
