% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{generateKeyFile}
\alias{generateKeyFile}
\title{Generate a Decryption Key File}
\usage{
generateKeyFile(target)
}
\arguments{
\item{target}{full path to the file to which the key shall be written}
}
\description{
Generate a Decryption Key File
}
