% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homedir.R
\name{get_homedir}
\alias{get_homedir}
\title{Get file path of user's home directory}
\usage{
get_homedir()
}
\value{
path of user's home directory on windows (full path 
home directory path but without '/Documents') or linux
}
\description{
Get file path of user's home directory
}
