% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homedir.R
\name{get_homedir}
\alias{get_homedir}
\title{Get File Path of User's Home Directory}
\usage{
get_homedir()
}
\value{
path of user's home directory on windows (full path home directory
  path but without '/Documents') or linux
}
\description{
Get File Path of User's Home Directory
}
