% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute.R
\name{getAttribute}
\alias{getAttribute}
\title{Safely get the Attribute of an Object}
\usage{
getAttribute(x, attributeName, do.stop = TRUE)
}
\arguments{
\item{x}{object from which to take the attribute}

\item{attributeName}{name of the attribute to be returned}

\item{do.stop}{if \code{TRUE} (default) an error is raised if the attribute
does not exist.}
}
\description{
Safely get the attribute of an object. An error is given if the attribute 
does not exist (and do.stop = TRUE)
}
\examples{

x <- structure(1, a = 2)
# getAttribute(x, "b") # gives a clear error message
identical(getAttribute(x, "a"), attr(x, "a")) # is TRUE

# Get an attribute's attribute by means of a "path" notation
y <- structure(1, a = structure(2, b = 3))
z <- structure(4, y = y)

str(y)
str(z)

kwb.utils::getAttribute(y, "a/b")
kwb.utils::getAttribute(z, "y/a/b")

}
