% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getFunctionName}
\alias{getFunctionName}
\title{Get the Name of a Function}
\usage{
getFunctionName(FUN)
}
\arguments{
\item{FUN}{R object representing a function}
}
\description{
Get the Name of a Function
}
