% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{getNamesOfObjectsInRDataFiles}
\alias{getNamesOfObjectsInRDataFiles}
\title{Deprecated. Use \code{\link{listObjects}} instead.}
\usage{
getNamesOfObjectsInRDataFiles(files.rdata)
}
\arguments{
\item{files.rdata}{vector of full paths to .RData files}
}
\description{
Deprecated. Use \code{\link{listObjects}} instead.
}
