% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{getPassword}
\alias{getPassword}
\title{Get Encrypted Password from File Using Key}
\usage{
getPassword(passwordFile, keyFile)
}
\arguments{
\item{passwordFile}{path to the password file}

\item{keyFile}{path to the key file}
}
\value{
NA if no password is stored
}
\description{
Get Encrypted Password from File Using Key
}
