% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsEncryption.R
\name{getPassword}
\alias{getPassword}
\title{get encrypted password from file using key}
\usage{
getPassword(passwordFile, keyFile)
}
\arguments{
\item{passwordFile}{path to the password file}

\item{keyFile}{path to the key file}
}
\value{
NA if no password is stored
}
\description{
get encrypted password from file using key
}
