% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryption.R
\name{getPassword}
\alias{getPassword}
\title{Get Encrypted Password from File Using Key}
\usage{
getPassword(passwordFile, keyFile)
}
\arguments{
\item{passwordFile}{path to the password file, as generated by 
\code{\link{createPasswordFile}}.}

\item{keyFile}{path to the key file, as generated by 
\code{\link{generateKeyFile}}.}
}
\value{
The password (character) or \code{NA} if no password is stored, both
  invisibly.
}
\description{
Given the path to a file containing an encrypted password and given the path
to the file containing the key that was used to do the encryption this 
function returns the original password, invisibly.
}
\seealso{
\code{\link{generateKeyFile}}, \code{\link{createPasswordFile}}
}
