\name{guessSeparator}
\alias{guessSeparator}
\title{guess column separator from file}
\description{guess column separator from file}
\usage{guessSeparator(csvFile, n = 10, separators = c(";", ",", "\\t"))}
\arguments{
  \item{csvFile}{full path to text file containing 'comma separated values'}
  \item{n}{number of first lines in the file to be looked at}
  \item{separators}{
}
}



\author{Hauke Sonnenberg}





