% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{hsChrToNum}
\alias{hsChrToNum}
\title{hsChrToNum}
\usage{
hsChrToNum(x, country, stopOnError = TRUE)
}
\arguments{
\item{x}{(vector of) text value(s) to be converted to numeric}

\item{country}{"en" if value(s) in \emph{x} is (are)  given in English format (decimal
point ".", thousands separator ",") or "de" if value is given in German
format (decimal point ",", thousands separator ".").}

\item{stopOnError}{if TRUE (default) the program stops if any of the given values could not
be converted.}
}
\description{
conversion of text representing a number to a valid number
}
