% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsString.R
\name{hsCountInStr}
\alias{hsCountInStr}
\title{Count Pattern in String}
\usage{
hsCountInStr(chr, str)
}
\arguments{
\item{chr}{character string (or pattern) to be looked for and counted in 
\code{str}}

\item{str}{character string in which to count for \code{chr}}
}
\value{
number of occurrences of \code{char} in \code{str}
}
\description{
Count occurrences of \code{chr} in \code{str}
}
