% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{hsDelEmptyCols}
\alias{hsDelEmptyCols}
\title{Delete empty Columns of Data Frame}
\usage{
hsDelEmptyCols(dataFrame, FUN = function(x) all(is.na(x)), drop = FALSE)
}
\arguments{
\item{dataFrame}{data frame of which empty columns (NA in all rows) are to be removed}

\item{FUN}{function to be applied to each column to decide whether the column
is empty or not. Default: \code{function(x) all(is.na(x))}}

\item{drop}{if \code{TRUE} (the default is \code{FALSE}) one dimension is 
dropped (a vector is returned instead of a data frame) in case that all but
 one columns are removed.}
}
\value{
copy of input data frame but with all empty columns removed
}
\description{
Returns data frame in which all empty columns (NA in all rows) are removed
}
\seealso{
\code{\link{removeEmptyColumns}}
}
