% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{hsPrepPdf}
\alias{hsPrepPdf}
\title{Prepare Writing of PDF File}
\usage{
hsPrepPdf(
  strPdf = tempfile(fileext = ".pdf"),
  boolLandscape = TRUE,
  bordW = 2,
  bordH = 2,
  makeCur = TRUE,
  ...
)
}
\arguments{
\item{strPdf}{see argument \code{pdfFile} in \code{\link{preparePdf}}}

\item{boolLandscape}{see argument \code{landscape} in \code{\link{preparePdf}}}

\item{bordW}{see argument \code{borderWidth.cm} in \code{\link{preparePdf}}}

\item{bordH}{see argument \code{borderHeight.cm} in \code{\link{preparePdf}}}

\item{makeCur}{see argument \code{makeCurrent} in \code{\link{preparePdf}}}

\item{\dots}{see \dots in \code{\link{preparePdf}}}
}
\description{
Deprecated. Please use \code{\link{preparePdf}} instead.
}
