\name{hsQuoteChr}
\alias{hsQuoteChr}
\title{hsQuoteChr}
\description{quotes objects of type character with quoting character}
\usage{hsQuoteChr(x, qchar = "'", escapeMethod = c("double", "backslash", 
    "none"))}
\arguments{
  \item{x}{vector or list of elements}
  \item{qchar}{quoting character to be used. Default: single quote "'"}
  \item{escapeMethod}{
}
}



\author{Hauke Sonnenberg}





