% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsMain.R
\name{hsRestoreAttributes}
\alias{hsRestoreAttributes}
\title{Restores object attributes}
\usage{
hsRestoreAttributes(x, attribs)
}
\arguments{
\item{x}{object}

\item{attribs}{former attributes of x (as retrieved by attributes(x)) to be restored}
}
\description{
Restores given attributes that are not object attributes any more
}
