% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute.R
\name{hsRestoreAttributes}
\alias{hsRestoreAttributes}
\title{Restore Object Attributes}
\usage{
hsRestoreAttributes(x, attribs)
}
\arguments{
\item{x}{object}

\item{attribs}{former attributes of x (as retrieved by attributes(x)) to be
restored}
}
\description{
Restore given attributes that are not object attributes any more
}
