% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{hsSafeName}
\alias{hsSafeName}
\title{Non-existing desired name}
\usage{
hsSafeName(myName, myNames)
}
\arguments{
\item{myName}{desired name.}

\item{myNames}{vector of existing names.}
}
\value{
If \emph{myName} is not contained in \emph{myNames} it is returned.
  Otherwise \emph{myName} is modified to \emph{myName}_01, \emph{myName}_02,
  ... until a non-existing name is found that is then returned.
}
\description{
Returns a name that is not yet contained in a vector \emph{myNames}
  of existing names.
}
\examples{
  existing <- c("a", "b")
  myName <- hsSafeName("c", existing)
  myName                               # "c"
  myName <- hsSafeName("a", existing)
  myName                               # "a_1"
  hsSafeName("a", c(existing, myName)) # "a_2"
  

}
