\name{hsStringToDouble}
\alias{hsStringToDouble}
\title{convert string to double}
\description{convert string to double considering given decimal sign in input string}
\usage{hsStringToDouble(strDbl, dec = ".")}
\arguments{
  \item{strDbl}{
}
  \item{dec}{
}
}

\value{double representation of input string \emph{strDbl}}

\author{Hauke Sonnenberg}





