% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsConvert.R
\name{hsStringToDouble}
\alias{hsStringToDouble}
\title{Convert String to Double}
\usage{
hsStringToDouble(strDbl, dec = ".")
}
\arguments{
\item{strDbl}{character string representing a double value}

\item{dec}{decimal character (default: \code{"."})}
}
\value{
double representation of input string \code{strDbl}
}
\description{
convert string to double considering given decimal sign in input string
}
