\name{hsSubstSpecChars}
\alias{hsSubstSpecChars}
\title{hsSubstSpecChars}
\description{Substitution of special characters}
\usage{hsSubstSpecChars(x)}
\arguments{
  \item{x}{string containing special characters to be substituted}
}

\value{input string \emph{x} with special characters being substituted by 
a meaningful represenation or underscore, multiple underscores replaced
by a single underscore and multiple underscores at the end removed.}

\author{Hauke Sonnenberg}





