% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{hsSubstSpecChars}
\alias{hsSubstSpecChars}
\title{Substitution of Special Characters}
\usage{
hsSubstSpecChars(x)
}
\arguments{
\item{x}{string containing special characters to be substituted}
}
\value{
input string \emph{x} with special characters being substituted by 
  a meaningful represenation or underscore, multiple underscores replaced
  by a single underscore and multiple underscores at the end removed.
}
\description{
Substitution of Special Characters
}
