% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{hsValidValue}
\alias{hsValidValue}
\title{Value in Correct English/German Notation?}
\usage{
hsValidValue(x, lng, dbg = FALSE, accept.na = TRUE)
}
\arguments{
\item{x}{vector of character strings}

\item{lng}{language code: "en" for English or "de" for German}

\item{dbg}{if \code{TRUE} (the default is \code{FALSE}) debug messages are
shown}

\item{accept.na}{if \code{TRUE} (default) \code{TRUE} is returned for 
\code{NA} values within \code{x}}
}
\description{
Returns TRUE if text representation of number is in correct format in 
  terms of decimal character and (optionally) thousand's separator character.
}
