% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{inRange}
\alias{inRange}
\title{check for values within minimum and maximum value}
\usage{
inRange(values, min.value, max.value, rng = NULL)
}
\arguments{
\item{values}{vector of values}

\item{min.value}{minimum value (inclusive)}

\item{max.value}{maximum value (inclusive)}

\item{rng}{optional. Vector of two elements representing \code{min.value}
and \code{max.value}, respectively}
}
\value{
vector of boolean
}
\description{
check for values within minimum and maximum value
}
