% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsLogical.R
\name{isNaInAllColumns}
\alias{isNaInAllColumns}
\title{isNaInAllColumns}
\usage{
isNaInAllColumns(dataFrame)
}
\arguments{
\item{dataFrame}{data frame or matrix}
}
\value{
logical vector with as many elements as there are rows in 
  \emph{dataFrame} (TRUE for rows in which all elements are NA, FALSE for
  rows in which there is at least one non-NA element).
}
\description{
isNaInAllColumns
}
