\name{isNaInAllRows}
\alias{isNaInAllRows}
\title{isNaInAllRows}
\description{isNaInAllRows}
\usage{isNaInAllRows(dataFrame)}
\arguments{
  \item{dataFrame}{data frame or matrix}
}

\value{logical vector with as many elements as there are columns in 
\emph{dataFrame} (TRUE for columns in which all elements are NA, FALSE for
columns in which there is at least one non-NA element).}

\author{Hauke Sonnenberg}





