% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{isNaOrEmpty}
\alias{isNaOrEmpty}
\title{NA or the empty string ""?}
\usage{
isNaOrEmpty(x)
}
\arguments{
\item{x}{object to be tested for NA or being empty (equal to "", after trimming)}
}
\value{
(vector of) logical, being TRUE for each element in \emph{x} that is
  NA or the empty string "" (after trimming)
}
\description{
is an object NA or equal to the empty string "" (after trimming)?
}
