% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical.R
\name{isNullOrEmpty}
\alias{isNullOrEmpty}
\title{isNullOrEmpty}
\usage{
isNullOrEmpty(x)
}
\arguments{
\item{x}{object to be tested for NULL or being empty (vector or list of
length 0 or data frame with no rows)}
}
\value{
TRUE if x is NULL or x is a vector of length 0 or x is a data frame
  with no rows.
}
\description{
isNullOrEmpty
}
