% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsMain.R
\name{lastElement}
\alias{lastElement}
\title{last element}
\usage{
lastElement(x)
}
\arguments{
\item{x}{object}
}
\value{
last element: x[length(x)]
}
\description{
Returns the last element using the function tail
}
