% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{makeUnique}
\alias{makeUnique}
\title{Make Duplicated Character Strings Unique
 
adds ".1", ".2", etc. to duplicate values}
\usage{
makeUnique(x, warn = TRUE)
}
\arguments{
\item{x}{vector of character strings}

\item{warn}{if \code{TRUE} (default) a warning showing the duplicated values is 
given}
}
\value{
\code{x} with duplicate elements being modified to "element.1",
  "element.2", etc.
}
\description{
Make Duplicated Character Strings Unique
 
adds ".1", ".2", etc. to duplicate values
}
