\name{makeUnique}
\alias{makeUnique}
\title{adds ".1", ".2", etc. to duplicate values}
\description{# adds ".1", ".2", etc. to duplicate values}
\usage{makeUnique(x, warn = TRUE)}
\arguments{
  \item{x}{vector of values}
  \item{warn}{
}
}

\value{\emph{x} with duplicate elements being modified to "element.1",
"element.2", etc.}

\author{Hauke Sonnenberg}





