% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{moveColumnsToFront}
\alias{moveColumnsToFront}
\title{move columns to the start of a data frame}
\usage{
moveColumnsToFront(x, columns = NULL)
}
\arguments{
\item{x}{data frame}

\item{columns}{vector of column names}
}
\value{
data frame or matrix with \code{columns} being the leftmost columns
}
\description{
move columns to the start of a data frame or matrix
}
\examples{
  x <- data.frame(a = 1:5, b = 2:6, c = 3:7)
  
  moveColumnsToFront(x, "b")
  moveColumnsToFront(x, c("b", "a"))
  

}
