% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsString.R
\name{multiSubstitute}
\alias{multiSubstitute}
\title{Multiple Substitutions}
\usage{
multiSubstitute(strings, replacements, ..., dbg = FALSE)
}
\arguments{
\item{strings}{vector of character}

\item{replacements}{list of pattern = replacement pairs.}

\item{\dots}{additional arguments passed to gsub}

\item{dbg}{if \code{TRUE} (the default is \code{FALSE}) it is shown which
strings were replaced}
}
\description{
apply multiple substitutions on a vector of character. For each element in
  \emph{replacements} gsub is called with the element name being the pattern
  and the element value being the replacement.
}
