% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{naToLastNonNa}
\alias{naToLastNonNa}
\title{Replace NA With "Last" non-NA}
\usage{
naToLastNonNa(x, method = 2)
}
\arguments{
\item{x}{vector in which NA are to be replaced with the last non-NA value 
(at greatest of smaller indices) in the vector}

\item{method}{integer (1 or 2) distinguishing two different methods}
}
\description{
replace NA values in a vector with the "last" non-NA values (at the nearest
  smaller indices in each case) in the vector
}
\examples{
  naToLastNonNa(c(1, 2, NA, NA, 3, NA, NA, 4, NA, NA, 5))
  ## Result: [1] 1 2 2 2 3 3 3 4 4 4 5
  
  # You will get an error if the first element is NA!
  
  # naToLastNonNa(c(NA, 1, NA, 2))
  
  ## Error in naToLastNonNa(c(NA, 1, NA, 2)) :
  ##   The first element must not be NA
  
}
