% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{orderBy}
\alias{orderBy}
\title{Order a Data Frame by One or more Columns}
\usage{
orderBy(df, by = NULL, ...)
}
\arguments{
\item{df}{data frame}

\item{by}{vector of column names specifying the columns by which to order}

\item{\dots}{further arguments passed to \code{\link{order}}, such as
\code{decreasing}}
}
\value{
\code{df} being sorted and with newly renumbered rows
}
\description{
Order a Data Frame by One or more Columns
}
\examples{
df <- data.frame(number = 4:1, letter = LETTERS[1:4])
df
orderBy(df, "number")
orderBy(df, "letter", decreasing = TRUE)

}
