% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsStat.R
\name{percentageOfMaximum}
\alias{percentageOfMaximum}
\title{percentageOfMaximum}
\usage{
percentageOfMaximum(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector of numeric values}

\item{na.rm}{passed to \code{max}}
}
\value{
100 * x / max(x)
}
\description{
percentageOfMaximum
}
