\name{percentageOfMaximum}
\alias{percentageOfMaximum}
\title{percentageOfMaximum}
\description{percentageOfMaximum}
\usage{percentageOfMaximum(x, na.rm = TRUE)}
\arguments{
  \item{x}{vector of numeric values}
  \item{na.rm}{passed to \code{max}}
}

\value{100 * x / max(x)}

\author{Hauke Sonnenberg}





