% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsStat.R
\name{percentageOfSum}
\alias{percentageOfSum}
\title{percentage of the sum of values}
\usage{
percentageOfSum(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector of numeric values}

\item{na.rm}{passed to \code{max}}
}
\value{
100 * x / sum(x)
}
\description{
percentage of the sum of values
}
\examples{
  p <- percentageOfSum(1:10)
  stopifnot(sum(p) == 100)
  

}
