% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{posixColumnAtPosition}
\alias{posixColumnAtPosition}
\title{Indices of POSIX columns in a Data Frame}
\usage{
posixColumnAtPosition(x)
}
\arguments{
\item{x}{data frame containing a date/time column}
}
\description{
Indices of POSIX columns in a Data Frame
}
