% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{preparePdf}
\alias{preparePdf}
\title{open PDF device with DIN A4 dimensions}
\usage{
preparePdf(pdfFile = tempfile(fileext = ".pdf"), landscape = TRUE,
  borderWidth.cm = 2, borderHeight.cm = 2,
  width.cm = .defaultWidth(landscape, borderWidth.cm, borderHeight.cm),
  height.cm = .defaultHeight(landscape, borderWidth.cm, borderHeight.cm),
  makeCurrent = TRUE, ...)
}
\arguments{
\item{pdfFile}{Full path to PDF file to be created}

\item{landscape}{If TRUE (default), orientation in PDF file will be landscape, else
portrait}

\item{borderWidth.cm}{(Total) border width in "width" direction in cm}

\item{borderHeight.cm}{(Total) border width in "height" direction in cm}

\item{width.cm}{page width in cm. Default according to DIN A4}

\item{height.cm}{page height in cm. Default according to DIN A4}

\item{makeCurrent}{if TRUE (default), the opened PDF device will become the current device}

\item{\dots}{further arguments passed to \code{pdf}}
}
\value{
full path to pdf file
}
\description{
open PDF device with DIN A4 dimensions
}
