% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{preparePdf}
\alias{preparePdf}
\title{Open PDF Device with DIN A4 Dimensions by Default}
\usage{
preparePdf(pdfFile = tempfile(fileext = ".pdf"), landscape = TRUE,
  borderWidth.cm = 2, borderHeight.cm = 2,
  width.cm = .defaultWidth(landscape, borderWidth.cm, borderHeight.cm),
  height.cm = .defaultHeight(landscape, borderWidth.cm, borderHeight.cm),
  makeCurrent = TRUE, paper = paste0("a4", ifelse(landscape, "r", "")),
  ...)
}
\arguments{
\item{pdfFile}{Full path to PDF file to be created}

\item{landscape}{If \code{TRUE} (default), orientation in PDF file will be landscape, else 
portrait}

\item{borderWidth.cm}{(Total) border width in "width" direction in cm}

\item{borderHeight.cm}{(Total) border width in "height" direction in cm}

\item{width.cm}{page width in cm. Default according to DIN A4}

\item{height.cm}{page height in cm. Default according to DIN A4}

\item{makeCurrent}{if \code{TRUE} (default), the opened PDF device will
become the current device, otherwise the current device will be restored}

\item{paper}{passed to \code{\link[grDevices]{pdf}}. By default "A4" (if
\code{landscape = FALSE}) or "A4r" (if \code{landscape = TRUE}). Use
\code{paper = "special"} to use the dimensions of the plot.}

\item{\dots}{further arguments passed to \code{\link[grDevices]{pdf}}}
}
\value{
full path to pdf file
}
\description{
Opens a PDF device in A4 paper format. After calling this function all plots
go into the specified PDF file in \code{pdfFile}. Important: The PDF file
needs to be closed explicitely with grDevices::dev.off() after all desired
plots have been made.
}
\examples{
\dontrun{
# Open PDF file by giving a path to preparePdf(). The path is returned.
pdf_file <- preparePdf(file.path(tempdir(), "example_preparePdf.pdf"))

# Plot something
plot(x <- seq(-pi,pi,pi/100), sin(x), type = "l")

# Open PDF file in viewer
finishAndShowPdf(pdf_file)
}

}
\seealso{
\code{\link{finishAndShowPdf}}
}
