\name{preparePdfIf}
\alias{preparePdfIf}
\title{preparePdfIf}
\description{prepare pdf file if condition is met}
\usage{preparePdfIf(to.pdf, PDF = "", ...)}
\arguments{
  \item{to.pdf}{
}
  \item{PDF}{
}
  \item{\dots}{
}
}

\value{full path to pdf file created if condition is met or "" else}

\author{Hauke Sonnenberg}





