\name{printIf}
\alias{printIf}
\title{call print if condition is met}
\description{call print if condition is met}
\usage{printIf(condition, x, caption = "")}
\arguments{
  \item{condition}{if TRUE, print is called, else not}
  \item{x}{object to be printed}
  \item{caption}{optional. Caption line to be printed with cat before printing \emph{x}}
}



\author{Hauke Sonnenberg}





