% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{printIf}
\alias{printIf}
\title{Call Print If Condition Is Met}
\usage{
printIf(condition, x, caption = deparse(substitute(x)))
}
\arguments{
\item{condition}{if TRUE, print is called, else not}

\item{x}{object to be printed}

\item{caption}{optional. Caption line to be printed with cat before printing
\emph{x}}
}
\description{
Call Print If Condition Is Met
}
