% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{quotient}
\alias{quotient}
\title{Quotient}
\usage{
quotient(dividend, divisor, substitute.value = Inf, warn = TRUE)
}
\arguments{
\item{dividend}{number to be devided}

\item{divisor}{number by which dividend is to be devided}

\item{substitute.value}{value to be returned if divisor is 0}

\item{warn}{if TRUE, a warning is given if the divisor is zero}
}
\value{
quotient of dividend and divisor: dividend/divisor
}
\description{
Calculate the quotient of two numbers
}
