% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{readDictionary}
\alias{readDictionary}
\title{Read Dictionary from Text File}
\usage{
readDictionary(file, sorted = TRUE)
}
\arguments{
\item{file}{full path to dictionary file or connection as e.g. opened with
\code{file}}

\item{sorted}{if TRUE (default) the entries in the dictionary will be sorted
by their keys}
}
\description{
Reads a dictionary (a list of "key = value"-pairs) from a text file.
}
\examples{
file <- system.file("extdata", "dictionary.txt", package = "kwb.utils")

dictionary <- readDictionary(file)

resolve("file.out", dictionary, extension = "csv")
resolve("file.out", dictionary, extension = "pdf")

}
\seealso{
\code{\link{readDictionaries}}
}
