\name{recursiveNames}
\alias{recursiveNames}
\title{names of all sublists of a list}
\description{returns the names of all sublists of \emph{x} in the "$"-notation, e.g.
list$sublist$subsublist$subsubsublist}
\usage{recursiveNames(x, basename = "")}
\arguments{
  \item{x}{R list.}
  \item{basename}{name to be used as prefix for all names found. Default: ""}
}



\author{Hauke Sonnenberg}





