% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{recycle}
\alias{recycle}
\title{"Recycle" Vector to given Length}
\usage{
recycle(x, n)
}
\arguments{
\item{x}{vector to be "recycled"}

\item{n}{target length}
}
\description{
"Recycle" Vector to given Length
}
