\name{relativeCumulatedSum}
\alias{relativeCumulatedSum}
\title{relativeCumulatedSum}
\description{relative cumulated sum of a vector of values}
\usage{relativeCumulatedSum(values)}
\arguments{
  \item{values}{vector of numeric values}
}



\author{Hauke Sonnenberg}





