% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{removeAttributes}
\alias{removeAttributes}
\title{Returns object without attributes}
\usage{
removeAttributes(x)
}
\arguments{
\item{x}{object}
}
\value{
\emph{x}, but with its attributes removed
}
\description{
Returns object without attributes
}
