% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{removeAttributes}
\alias{removeAttributes}
\title{Returns object without attributes}
\usage{
removeAttributes(x, names = NULL)
}
\arguments{
\item{x}{object}

\item{names}{names of attributes to be removed. If \code{NULL} (default),
all attributes are removed.}
}
\value{
\emph{x}, but with its attributes removed
}
\description{
Returns object without attributes
}
