% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{removeColumns}
\alias{removeColumns}
\title{remove columns from data frame}
\usage{
removeColumns(dframe, columnsToRemove, drop = FALSE)
}
\arguments{
\item{dframe}{data frame,}

\item{columnsToRemove}{vector of column names giving the columns to remove}

\item{drop}{if FALSE, a data frame is returned in any case, otherwise the result may
be a vector if only one column remains}
}
\value{
\emph{dframe} with columns given in \emph{columnsToRemove} being removed.
  User attributes of \emph{dframe} are restored.
}
\description{
remove columns from a data frame
}
