% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{removeColumns}
\alias{removeColumns}
\title{Remove Columns from a Data Frame}
\usage{
removeColumns(
  dframe,
  columns = NULL,
  columnsToRemove = NULL,
  pattern = NULL,
  drop = FALSE,
  dbg = FALSE
)
}
\arguments{
\item{dframe}{data frame,}

\item{columns}{vector of column names giving the columns to remove}

\item{columnsToRemove}{deprecated. Use argument \code{columns} instead.}

\item{pattern}{regular expression matching the names of the columns to be
removed. Will only be evaluated if no explicit column names are given in 
\code{columns}.}

\item{drop}{if FALSE, a data frame is returned in any case, otherwise the
result may be a vector if only one column remains}

\item{dbg}{if \code{TRUE} (the default is \code{FALSE}), the deletion of
columns is reported on the screen}
}
\value{
\emph{dframe} with columns given in \emph{columns} being removed.
  User attributes of \emph{dframe} are restored.
}
\description{
Remove Columns from a Data Frame
}
