\name{removeSpaces}
\alias{removeSpaces}
\title{remove all spaces in string(s)}
\description{remove all spaces in string(s)}
\usage{removeSpaces(x)}
\arguments{
  \item{x}{(vector of) character}
}

\value{\emph{x} with all spaces removed}

\author{Hauke Sonnenberg}





