% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{removeSpaces}
\alias{removeSpaces}
\title{Remove all Spaces in String(s)}
\usage{
removeSpaces(x)
}
\arguments{
\item{x}{(vector of) character}
}
\value{
\emph{x} with all spaces removed
}
\description{
Remove all Spaces in String(s)
}
