% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column.R
\name{renameAndSelect}
\alias{renameAndSelect}
\title{Rename and Select Columns of a Data Frame}
\usage{
renameAndSelect(data, renames, columns = unlist(renames))
}
\arguments{
\item{data}{data frame}

\item{renames}{list defining renames in the form of "oldName" = "newName"
pairs}

\item{columns}{(new) names of columns to be selected}
}
\description{
Rename and Select Columns of a Data Frame
}
