% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{rStylePath}
\alias{rStylePath}
\title{R compatible file path}
\usage{
rStylePath(path)
}
\arguments{
\item{path}{character string representing a file path}
}
\value{
path in which backslashes are replaced with forward slashes
}
\description{
R compatible file path with backslashes replaced with forward slashes
}
