% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsSystem.R
\name{runBatchfileInDirectory}
\alias{runBatchfileInDirectory}
\title{runBatchfileInDirectory}
\usage{
runBatchfileInDirectory(batchfile, directory = dirname(batchfile), ...)
}
\arguments{
\item{batchfile}{full path to Windows batch file}

\item{directory}{directory from which batchfile is to be invoked. Default: directory
of batch file}

\item{\dots}{arguments passed to shell.exec}
}
\description{
runBatchfileInDirectory
}
