% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{runInDirectory}
\alias{runInDirectory}
\title{change working dir and run function}
\usage{
runInDirectory(target.dir = tempdir(), FUN, ..., .dbg = FALSE)
}
\arguments{
\item{target.dir}{target directory. Default: tempdir()}

\item{FUN}{function to be invoked}

\item{\dots}{arguments to be passed to function FUN}

\item{.dbg}{if TRUE, debug messages on changing the directory are shown. Default:
\code{FALSE}}
}
\description{
change working directory and run function
}
