% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsDataFrame.R
\name{safeColumnBind}
\alias{safeColumnBind}
\title{"Safe" version of cbind.}
\usage{
safeColumnBind(x1, x2)
}
\arguments{
\item{x1}{first object to be passed to \code{cbind}}

\item{x2}{second object to be passed to  \code{cbind}}
}
\value{
result of \code{cbind(x1, x2)} or \code{x2} if \code{x1}
  is \code{NULL}.
}
\description{
If \code{x1} is NULL \code{x2} is returned otherwise \code{cbind(x1, x2)}
}
\examples{
  x1 <- NULL
  
  for (i in 1:3) {
  
    x2 <- data.frame(a = 1:3, b = rnorm(3))
    x1 <- safeColumnBind(x1, x2)
  
    # using cbind would result in an error:
    # x1 <- cbind(x1, x2)
  }
  
  x1
  
}
