% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsSystem.R
\name{safePath}
\alias{safePath}
\title{stop if path does not exit}
\usage{
safePath(...)
}
\arguments{
\item{\dots}{character vectors passed to \code{file.path}}
}
\value{
path as given in \code{path}
}
\description{
stop if path does not exit and return the path
}
