% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataFrame.R
\name{safeRowBind}
\alias{safeRowBind}
\title{"safe" rbind}
\usage{
safeRowBind(dataFrame1, dataFrame2)
}
\arguments{
\item{dataFrame1}{first data frame}

\item{dataFrame2}{second data frame}
}
\description{
rbind two data frames even if column names differ
}
\examples{
kwb.utils::safeRowBind(
  data.frame(A = 1:2, B = 2:3),
  data.frame(B = 3:4, C = 4:5)
)
  
}
