% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{setMatrixColumns}
\alias{setMatrixColumns}
\title{Set Matrix Columns to Values}
\usage{
setMatrixColumns(m, columnValuePairs, warn = TRUE)
}
\arguments{
\item{m}{matrix}

\item{columnValuePairs}{list of elements each of which defines an assignment
in the form \code{<column-name> = <value>}}

\item{warn}{if \code{TRUE}, warnings are given if columns named in
\code{columnValuePairs} do not exist in matrix \code{m}}
}
\description{
Set matrix columns of given names to fix values
}
