% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{stringContains}
\alias{stringContains}
\title{stringContains}
\usage{
stringContains(x, contains)
}
\arguments{
\item{x}{vector of character}

\item{contains}{vector of character}
}
\description{
stringContains
}
\examples{
stringContains(c("abc", "Kabeljau", "Arabella"), "ab")
stringContains(c("abc", "Kabeljau", "Arabella"), "abc")
  
}
