% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string.R
\name{stringStartsWith}
\alias{stringStartsWith}
\title{stringStartsWith}
\usage{
stringStartsWith(x, startsWith)
}
\arguments{
\item{x}{vector of character to be checked if they start with
\code{startsWith}}

\item{startsWith}{string to be searched for at the beginning of the string(s)
in \code{x}}
}
\description{
stringStartsWith
}
\examples{
stringStartsWith(c("abc", "Kabeljau", "Arabella"), "ab")
stringStartsWith(c("abc", "Kabeljau", "Arabella"), "A")
  
}
