% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsSystem.R
\name{tempSubdirectory}
\alias{tempSubdirectory}
\title{create and return path of subdirectory in temp()}
\usage{
tempSubdirectory(subdir)
}
\arguments{
\item{subdir}{name of subdirectory to be created}
}
\value{
full path to created directory
}
\description{
create and return path of subdirectory in temp()
}
