% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{tempSubdirectory}
\alias{tempSubdirectory}
\title{Create and return Path of Subdirectory in temp()}
\usage{
tempSubdirectory(subdir)
}
\arguments{
\item{subdir}{name of subdirectory to be created}
}
\value{
full path to created directory
}
\description{
Create and return path of subdirectory in temp()
}
