\name{toInches}
\alias{toInches}
\title{convert cm to inches}
\description{# convert cm to inches}
\usage{toInches(cm)}
\arguments{
  \item{cm}{vector of numeric representing length(s) in cm}
}

\value{vector of numeric representing length(s) in inches}

\author{Hauke Sonnenberg}





