% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{toInches}
\alias{toInches}
\title{convert cm to inches}
\usage{
toInches(cm)
}
\arguments{
\item{cm}{vector of numeric representing length(s) in cm}
}
\value{
vector of numeric representing length(s) in inches
}
\description{
# convert cm to inches
}
