% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsMain.R
\name{warnIfEmpty}
\alias{warnIfEmpty}
\title{warnIfEmpty}
\usage{
warnIfEmpty(x)
}
\arguments{
\item{x}{object to be tested for NULL or being empty (vector of length 0 or
data frame with no rows)}
}
\description{
Gives a warning if the object is NULL or empty and returns the object
}
