% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{warningDeprecated}
\alias{warningDeprecated}
\title{Create Warning About a Deprecated Function}
\usage{
warningDeprecated(old_name, new_name, parentheses = TRUE)
}
\arguments{
\item{old_name}{name of deprecated function}

\item{new_name}{name of new function to be used instead}

\item{parentheses}{logical. Should parentheses be printed after the names
in the output? Default: \code{TRUE}.}
}
\description{
Create Warning About a Deprecated Function
}
\examples{
\dontrun{
warningDeprecated("old_function()", "new_function()")
}
}
