% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system.R
\name{windowsPath}
\alias{windowsPath}
\title{convert to MS Windows-compatible path}
\usage{
windowsPath(path)
}
\arguments{
\item{path}{vector of character representing file paths}
}
\description{
create MS Windows-compatible path by substituting forward slashes with 
  backslashes
}
